SwaggerPetstore.API

Description

Operations

addPet

addPet

Arguments

:: Pet
"body" - Pet object that needs to be added to the store

-> SwaggerPetstoreRequest ()

POST /pet

Add a new pet to the store

AuthMethod: petstore_auth

Consumes: applicationjson, applicationxml

Produces: applicationxml, applicationjson

deletePet

deletePet

Arguments

:: Integer
"petId" - Pet id to delete

-> SwaggerPetstoreRequest ()

DELETE /pet/{petId}

Deletes a pet

AuthMethod: petstore_auth

Produces: applicationxml, applicationjson

Optional Params:

"api_key" :: Text

findPetsByStatus

findPetsByStatus

Arguments

:: [Text]
"status" - Status values that need to be considered for filter

-> SwaggerPetstoreRequest [Pet]

GET /pet/findByStatus

Finds Pets by status

Multiple status values can be provided with comma separated strings

AuthMethod: petstore_auth

Produces: applicationxml, applicationjson

findPetsByTags

findPetsByTags

Arguments

:: [Text]
"tags" - Tags to filter by

-> SwaggerPetstoreRequest [Pet]

Deprecated:

GET /pet/findByTags

Finds Pets by tags

Muliple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.

AuthMethod: petstore_auth

Produces: applicationxml, applicationjson

getPetById

getPetById

Arguments

:: Integer
"petId" - ID of pet to return

-> SwaggerPetstoreRequest Pet

GET /pet/{petId}

Find pet by ID

Returns a single pet

AuthMethod: api_key

Produces: applicationxml, applicationjson

updatePet

updatePet

Arguments

:: Pet
"body" - Pet object that needs to be added to the store

-> SwaggerPetstoreRequest ()

PUT /pet

Update an existing pet

AuthMethod: petstore_auth

Consumes: applicationjson, applicationxml

Produces: applicationxml, applicationjson

updatePetWithForm

updatePetWithForm

Arguments

:: Integer
"petId" - ID of pet that needs to be updated

-> SwaggerPetstoreRequest ()

POST /pet/{petId}

Updates a pet in the store with form data

AuthMethod: petstore_auth

Consumes: application/x-www-form-urlencoded

Produces: applicationxml, applicationjson

Optional Params:

"name" :: Text - Updated name of the pet

"status" :: Text - Updated status of the pet

uploadFile

uploadFile

Arguments

:: Integer
"petId" - ID of pet to update

-> SwaggerPetstoreRequest ApiResponse

POST /pet/{petId}/uploadImage

uploads an image

AuthMethod: petstore_auth

Consumes: multipart/form-data

Produces: application/json

Optional Params:

"additionalMetadata" :: Text - Additional data to pass to server

"file" :: FilePath - file to upload

deleteOrder

deleteOrder

Arguments

:: Integer
"orderId" - ID of the order that needs to be deleted

-> SwaggerPetstoreRequest ()

DELETE /store/order/{orderId}

Delete purchase order by ID

For valid response try integer IDs with positive integer value. Negative or non-integer values will generate API errors

Produces: applicationxml, applicationjson

getInventory

getInventory :: SwaggerPetstoreRequest (Map String Int)

GET /store/inventory

Returns pet inventories by status

Returns a map of status codes to quantities

AuthMethod: api_key

Produces: application/json

getOrderById

getOrderById

Arguments

:: Integer
"orderId" - ID of pet that needs to be fetched

-> SwaggerPetstoreRequest Order

GET /store/order/{orderId}

Find purchase order by ID

For valid response try integer IDs with value >= 1 and <= 10. Other values will generated exceptions

Produces: applicationxml, applicationjson

placeOrder

placeOrder

Arguments

:: Order
"body" - order placed for purchasing the pet

-> SwaggerPetstoreRequest Order

POST /store/order

Place an order for a pet

Produces: applicationxml, applicationjson

createUser

createUser

Arguments

:: User
"body" - Created user object

-> SwaggerPetstoreRequest ()

POST /user

Create user

This can only be done by the logged in user.

Produces: applicationxml, applicationjson

createUsersWithArrayInput

createUsersWithArrayInput

Arguments

:: [User]
"body" - List of user object

-> SwaggerPetstoreRequest ()

POST /user/createWithArray

Creates list of users with given input array

Produces: applicationxml, applicationjson

createUsersWithListInput

createUsersWithListInput

Arguments

:: [User]
"body" - List of user object

-> SwaggerPetstoreRequest ()

POST /user/createWithList

Creates list of users with given input array

Produces: applicationxml, applicationjson

deleteUser

deleteUser

Arguments

:: Text
"username" - The name that needs to be deleted

-> SwaggerPetstoreRequest ()

DELETE /user/{username}

Delete user

This can only be done by the logged in user.

Produces: applicationxml, applicationjson

getUserByName

getUserByName

Arguments

:: Text
"username" - The name that needs to be fetched. Use user1 for testing.

-> SwaggerPetstoreRequest User

GET /user/{username}

Get user by user name

Produces: applicationxml, applicationjson

loginUser

loginUser

Arguments

:: Text
"username" - The user name for login

-> Text
"password" - The password for login in clear text

-> SwaggerPetstoreRequest Text

GET /user/login

Logs user into the system

Produces: applicationxml, applicationjson

logoutUser

logoutUser :: SwaggerPetstoreRequest ()

GET /user/logout

Logs out current logged in user session

Produces: applicationxml, applicationjson

updateUser

updateUser

Arguments

:: Text
"username" - name that need to be updated

-> User
"body" - Updated user object

-> SwaggerPetstoreRequest ()

PUT /user/{username}

Updated user

This can only be done by the logged in user.

Produces: applicationxml, applicationjson

SwaggerPetstoreRequest

data SwaggerPetstoreRequest r

Constructors

SwaggerPetstoreRequest

Fields

rMethod :: Method

Method of SwaggerPetstoreRequest

endpoint :: [Text]

Endpoint of SwaggerPetstoreRequest

params :: [EncodedParam]

Encoded params of SwaggerPetstoreRequest

mkSwaggerPetstoreRequest :: Method -> [Text] -> [EncodedParam] -> SwaggerPetstoreRequest r

type TupleBS8 = (ByteString, ByteString)

Type alias for query parameters

EncodedParam

data EncodedParam

Constructors

Query TupleBS8

Body ByteString

Header TupleBS8

data ResultFormatType

Constructors

FormatJson

FormatXml

toPath :: IsString t => t1 -> t