SwaggerPetstore.API.Pet

Description

Operations

Pet

addPet

addPet

Arguments

:: (Consumes AddPet contentType, MimeRender contentType Pet)

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Pet
"body" - Pet object that needs to be added to the store

-> SwaggerPetstoreRequest AddPet contentType res accept

POST /pet

Add a new pet to the store

AuthMethod: AuthOAuthPetstoreAuth

Note: Has Produces instances, but no response schema

data AddPet

deletePet

deletePet

Arguments

:: Accept accept
request accept (MimeType)

-> PetId
"petId" - Pet id to delete

-> SwaggerPetstoreRequest DeletePet MimeNoContent res accept

DELETE /pet/{petId}

Deletes a pet

AuthMethod: AuthOAuthPetstoreAuth

Note: Has Produces instances, but no response schema

data DeletePet

findPetsByStatus

findPetsByStatus

Arguments

:: Accept accept
request accept (MimeType)

-> Status
"status" - Status values that need to be considered for filter

-> SwaggerPetstoreRequest FindPetsByStatus MimeNoContent [Pet] accept

GET /pet/findByStatus

Finds Pets by status

Multiple status values can be provided with comma separated strings

AuthMethod: AuthOAuthPetstoreAuth

data FindPetsByStatus

findPetsByTags

findPetsByTags

Arguments

:: Accept accept
request accept (MimeType)

-> Tags
"tags" - Tags to filter by

-> SwaggerPetstoreRequest FindPetsByTags MimeNoContent [Pet] accept

Deprecated:

GET /pet/findByTags

Finds Pets by tags

Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.

AuthMethod: AuthOAuthPetstoreAuth

data FindPetsByTags

getPetById

getPetById

Arguments

:: Accept accept
request accept (MimeType)

-> PetId
"petId" - ID of pet to return

-> SwaggerPetstoreRequest GetPetById MimeNoContent Pet accept

GET /pet/{petId}

Find pet by ID

Returns a single pet

AuthMethod: AuthApiKeyApiKey

data GetPetById

updatePet

updatePet

Arguments

:: (Consumes UpdatePet contentType, MimeRender contentType Pet)

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Pet
"body" - Pet object that needs to be added to the store

-> SwaggerPetstoreRequest UpdatePet contentType res accept

PUT /pet

Update an existing pet

AuthMethod: AuthOAuthPetstoreAuth

Note: Has Produces instances, but no response schema

data UpdatePet

updatePetWithForm

updatePetWithForm

Arguments

:: Consumes UpdatePetWithForm MimeFormUrlEncoded

=> Accept accept
request accept (MimeType)

-> PetId
"petId" - ID of pet that needs to be updated

-> SwaggerPetstoreRequest UpdatePetWithForm MimeFormUrlEncoded res accept

POST /pet/{petId}

Updates a pet in the store with form data

AuthMethod: AuthOAuthPetstoreAuth

Note: Has Produces instances, but no response schema

data UpdatePetWithForm

uploadFile

uploadFile

Arguments

:: Consumes UploadFile MimeMultipartFormData

=> PetId
"petId" - ID of pet to update

-> SwaggerPetstoreRequest UploadFile MimeMultipartFormData ApiResponse MimeJSON

POST /pet/{petId}/uploadImage

uploads an image

AuthMethod: AuthOAuthPetstoreAuth

data UploadFile