SwaggerPetstore.API.User

Description

Operations

User

createUser

createUser

Arguments

:: (Consumes CreateUser contentType, MimeRender contentType User)

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> User
"body" - Created user object

-> SwaggerPetstoreRequest CreateUser contentType res accept

POST /user

Create user

This can only be done by the logged in user.

Note: Has Produces instances, but no response schema

data CreateUser

createUsersWithArrayInput

createUsersWithArrayInput

Arguments

:: (Consumes CreateUsersWithArrayInput contentType, MimeRender contentType Body)

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Body
"body" - List of user object

-> SwaggerPetstoreRequest CreateUsersWithArrayInput contentType res accept

POST /user/createWithArray

Creates list of users with given input array

Note: Has Produces instances, but no response schema

data CreateUsersWithArrayInput

createUsersWithListInput

createUsersWithListInput

Arguments

:: (Consumes CreateUsersWithListInput contentType, MimeRender contentType Body)

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Body
"body" - List of user object

-> SwaggerPetstoreRequest CreateUsersWithListInput contentType res accept

POST /user/createWithList

Creates list of users with given input array

Note: Has Produces instances, but no response schema

data CreateUsersWithListInput

deleteUser

deleteUser

Arguments

:: Accept accept
request accept (MimeType)

-> Username
"username" - The name that needs to be deleted

-> SwaggerPetstoreRequest DeleteUser MimeNoContent res accept

DELETE /user/{username}

Delete user

This can only be done by the logged in user.

Note: Has Produces instances, but no response schema

data DeleteUser

getUserByName

getUserByName

Arguments

:: Accept accept
request accept (MimeType)

-> Username
"username" - The name that needs to be fetched. Use user1 for testing.

-> SwaggerPetstoreRequest GetUserByName MimeNoContent User accept

GET /user/{username}

Get user by user name

data GetUserByName

loginUser

loginUser

Arguments

:: Accept accept
request accept (MimeType)

-> Username
"username" - The user name for login

-> Password
"password" - The password for login in clear text

-> SwaggerPetstoreRequest LoginUser MimeNoContent Text accept

GET /user/login

Logs user into the system

data LoginUser

logoutUser

logoutUser

Arguments

:: Accept accept
request accept (MimeType)

-> SwaggerPetstoreRequest LogoutUser MimeNoContent res accept

GET /user/logout

Logs out current logged in user session

Note: Has Produces instances, but no response schema

data LogoutUser

updateUser

updateUser

Arguments

:: (Consumes UpdateUser contentType, MimeRender contentType User)

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Username
"username" - name that need to be deleted

-> User
"body" - Updated user object

-> SwaggerPetstoreRequest UpdateUser contentType res accept

PUT /user/{username}

Updated user

This can only be done by the logged in user.

Note: Has Produces instances, but no response schema

data UpdateUser