SwaggerPetstore.Client

Description

Config

data SwaggerPetstoreConfig

Constructors

SwaggerPetstoreConfig

Fields

configHost :: ByteString

host supplied in the Request

configUserAgent :: Text

user-agent supplied in the Request

configExecLoggingT :: ExecLoggingT

Run a block using a MonadLogger instance

configLoggingFilter :: LogSource -> LogLevel -> Bool

Only log messages passing the given predicate function.

newConfig :: SwaggerPetstoreConfig

constructs a default SwaggerPetstoreConfig

configHost:

http://petstore.swagger.io/v2

configUserAgent:

"swagger-haskell-http-client/1.0.0"

configExecLoggingT: runNullLoggingT

configLoggingFilter: infoLevelFilter

withStdoutLogging :: SwaggerPetstoreConfig -> SwaggerPetstoreConfig

updates the config to use a MonadLogger instance which prints to stdout.

withStderrLogging :: SwaggerPetstoreConfig -> SwaggerPetstoreConfig

updates the config to use a MonadLogger instance which prints to stderr.

withNoLogging :: SwaggerPetstoreConfig -> SwaggerPetstoreConfig

updates the config to disable logging

Dispatch

Lbs

dispatchLbs

Arguments

:: (Produces req accept, MimeType contentType)

=> Manager
http-client Connection manager

-> SwaggerPetstoreConfig
config

-> SwaggerPetstoreRequest req contentType res
request

-> accept
"accept" MimeType

-> IO (Response ByteString)
response

send a request returning the raw http response

Mime

data MimeResult res

pair of decoded http body and http response

Constructors

MimeResult

Fields

mimeResult :: Either MimeError res

decoded http body

mimeResultResponse :: Response ByteString

http response

data MimeError

pair of unrender/parser error and http response

Constructors

MimeError

Fields

mimeError :: String

unrender/parser error

mimeErrorResponse :: Response ByteString

http response

dispatchMime

Arguments

:: (Produces req accept, MimeUnrender accept res, MimeType contentType)

=> Manager
http-client Connection manager

-> SwaggerPetstoreConfig
config

-> SwaggerPetstoreRequest req contentType res
request

-> accept
"accept" MimeType

-> IO (MimeResult res)
response

send a request returning the MimeResult

dispatchMime'

Arguments

:: (Produces req accept, MimeUnrender accept res, MimeType contentType)

=> Manager
http-client Connection manager

-> SwaggerPetstoreConfig
config

-> SwaggerPetstoreRequest req contentType res
request

-> accept
"accept" MimeType

-> IO (Either MimeError res)
response

like dispatchMime, but only returns the decoded http body

Unsafe

dispatchLbsUnsafe

Arguments

:: (MimeType accept, MimeType contentType)

=> Manager
http-client Connection manager

-> SwaggerPetstoreConfig
config

-> SwaggerPetstoreRequest req contentType res
request

-> accept
"accept" MimeType

-> IO (Response ByteString)
response

like dispatchReqLbs, but does not validate the operation is a Producer of the "accept" MimeType. (Useful if the server's response is undocumented)

dispatchInitUnsafe

Arguments

:: Manager
http-client Connection manager

-> SwaggerPetstoreConfig
config

-> InitRequest req contentType res accept
init request

-> IO (Response ByteString)
response

dispatch an InitRequest

InitRequest

newtype InitRequest req contentType res accept

wraps an http-client Request with request/response type parameters

Constructors

InitRequest

Fields

unInitRequest :: Request

_toInitRequest

Arguments

:: (MimeType accept, MimeType contentType)

=> SwaggerPetstoreConfig
config

-> SwaggerPetstoreRequest req contentType res
request

-> accept
"accept" MimeType

-> IO (InitRequest req contentType res accept)
initialized request

Build an http-client Request record from the supplied config and request

modifyInitRequest :: InitRequest req contentType res accept -> (Request -> Request) -> InitRequest req contentType res accept

modify the underlying Request

modifyInitRequestM :: Monad m => InitRequest req contentType res accept -> (Request -> m Request) -> m (InitRequest req contentType res accept)

modify the underlying Request (monadic)

Logging

type ExecLoggingT = forall m. MonadIO m => forall a. LoggingT m a -> m a

A block using a MonadLogger instance

Null Logger

nullLogger :: Loc -> LogSource -> LogLevel -> LogStr -> IO ()

a logger which disables logging

runNullLoggingT :: LoggingT m a -> m a

run the monad transformer that disables logging

Logging Filters

errorLevelFilter :: LogSource -> LogLevel -> Bool

a log filter that uses LevelError as the minimum logging level

infoLevelFilter :: LogSource -> LogLevel -> Bool

a log filter that uses LevelInfo as the minimum logging level

debugLevelFilter :: LogSource -> LogLevel -> Bool

a log filter that uses LevelDebug as the minimum logging level

minLevelFilter :: LogLevel -> LogSource -> LogLevel -> Bool

Logging

logNST :: (MonadIO m, MonadLogger m) => LogLevel -> Text -> Text -> m ()

Log a message using the current time

logExceptions :: (MonadLogger m, MonadCatch m, MonadIO m) => Text -> m a -> m a

re-throws exceptions after logging them

runLoggingT :: SwaggerPetstoreConfig -> ExecLoggingT

Run a block using the configured MonadLogger instance

runExceptionLoggingT :: (MonadCatch m, MonadIO m) => Text -> SwaggerPetstoreConfig -> LoggingT m a -> m a

Run a block using the configured MonadLogger instance (logs exceptions)