SwaggerPetstore.Model

Description

Models

AdditionalPropertiesClass

data AdditionalPropertiesClass

AdditionalPropertiesClass

Constructors

AdditionalPropertiesClass

Fields

additionalPropertiesClassMapProperty :: !(Maybe (Map String Text))

"map_property"

additionalPropertiesClassMapOfMapProperty :: !(Maybe (Map String (Map String Text)))

"map_of_map_property"

mkAdditionalPropertiesClass :: AdditionalPropertiesClass

Construct a value of type AdditionalPropertiesClass (by applying it's required fields, if any)

Animal

data Animal

Animal

Constructors

Animal

Fields

animalClassName :: !Text

Required "className"

animalColor :: !(Maybe Text)

"color"

mkAnimal

Arguments

:: Text
animalClassName

-> Animal

Construct a value of type Animal (by applying it's required fields, if any)

AnimalFarm

data AnimalFarm

AnimalFarm

Constructors

AnimalFarm

mkAnimalFarm :: AnimalFarm

Construct a value of type AnimalFarm (by applying it's required fields, if any)

ApiResponse

data ApiResponse

ApiResponse

Constructors

ApiResponse

Fields

apiResponseCode :: !(Maybe Int)

"code"

apiResponseType :: !(Maybe Text)

"type"

apiResponseMessage :: !(Maybe Text)

"message"

mkApiResponse :: ApiResponse

Construct a value of type ApiResponse (by applying it's required fields, if any)

ArrayOfArrayOfNumberOnly

data ArrayOfArrayOfNumberOnly

ArrayOfArrayOfNumberOnly

Constructors

ArrayOfArrayOfNumberOnly

Fields

arrayOfArrayOfNumberOnlyArrayArrayNumber :: !(Maybe [[Double]])

ArrayArrayNumber

mkArrayOfArrayOfNumberOnly :: ArrayOfArrayOfNumberOnly

Construct a value of type ArrayOfArrayOfNumberOnly (by applying it's required fields, if any)

ArrayOfNumberOnly

data ArrayOfNumberOnly

ArrayOfNumberOnly

Constructors

ArrayOfNumberOnly

Fields

arrayOfNumberOnlyArrayNumber :: !(Maybe [Double])

ArrayNumber

mkArrayOfNumberOnly :: ArrayOfNumberOnly

Construct a value of type ArrayOfNumberOnly (by applying it's required fields, if any)

ArrayTest

data ArrayTest

ArrayTest

Constructors

ArrayTest

Fields

arrayTestArrayOfString :: !(Maybe [Text])

"array_of_string"

arrayTestArrayArrayOfInteger :: !(Maybe [[Integer]])

"array_array_of_integer"

arrayTestArrayArrayOfModel :: !(Maybe [[ReadOnlyFirst]])

"array_array_of_model"

mkArrayTest :: ArrayTest

Construct a value of type ArrayTest (by applying it's required fields, if any)

Capitalization

data Capitalization

Capitalization

Constructors

Capitalization

Fields

capitalizationSmallCamel :: !(Maybe Text)

"smallCamel"

capitalizationCapitalCamel :: !(Maybe Text)

CapitalCamel

capitalizationSmallSnake :: !(Maybe Text)

"small_Snake"

capitalizationCapitalSnake :: !(Maybe Text)

Capital_Snake

capitalizationScaEthFlowPoints :: !(Maybe Text)

SCA_ETH_Flow_Points

capitalizationAttName :: !(Maybe Text)

ATT_NAME - Name of the pet

mkCapitalization :: Capitalization

Construct a value of type Capitalization (by applying it's required fields, if any)

Category

data Category

Category

Constructors

Category

Fields

categoryId :: !(Maybe Integer)

"id"

categoryName :: !(Maybe Text)

"name"

mkCategory :: Category

Construct a value of type Category (by applying it's required fields, if any)

ClassModel

data ClassModel

ClassModel Model for testing model with "_class" property

Constructors

ClassModel

Fields

classModelClass :: !(Maybe Text)

"_class"

mkClassModel :: ClassModel

Construct a value of type ClassModel (by applying it's required fields, if any)

Client

data Client

Client

Constructors

Client

Fields

clientClient :: !(Maybe Text)

"client"

mkClient :: Client

Construct a value of type Client (by applying it's required fields, if any)

EnumArrays

data EnumArrays

EnumArrays

Constructors

EnumArrays

Fields

enumArraysJustSymbol :: !(Maybe Text)

"just_symbol"

enumArraysArrayEnum :: !(Maybe [Text])

"array_enum"

mkEnumArrays :: EnumArrays

Construct a value of type EnumArrays (by applying it's required fields, if any)

EnumClass

data EnumClass

EnumClass

Constructors

EnumClass

mkEnumClass :: EnumClass

Construct a value of type EnumClass (by applying it's required fields, if any)

EnumTest

data EnumTest

EnumTest

Constructors

EnumTest

Fields

enumTestEnumString :: !(Maybe Text)

"enum_string"

enumTestEnumInteger :: !(Maybe Int)

"enum_integer"

enumTestEnumNumber :: !(Maybe Double)

"enum_number"

enumTestOuterEnum :: !(Maybe OuterEnum)

"outerEnum"

mkEnumTest :: EnumTest

Construct a value of type EnumTest (by applying it's required fields, if any)

FormatTest

data FormatTest

FormatTest

Constructors

FormatTest

Fields

formatTestInteger :: !(Maybe Int)

"integer"

formatTestInt32 :: !(Maybe Int)

"int32"

formatTestInt64 :: !(Maybe Integer)

"int64"

formatTestNumber :: !Double

Required "number"

formatTestFloat :: !(Maybe Float)

"float"

formatTestDouble :: !(Maybe Double)

"double"

formatTestString :: !(Maybe Text)

"string"

formatTestByte :: !ByteArray

Required "byte"

formatTestBinary :: !(Maybe Binary)

"binary"

formatTestDate :: !Date

Required "date"

formatTestDateTime :: !(Maybe DateTime)

"dateTime"

formatTestUuid :: !(Maybe Text)

"uuid"

formatTestPassword :: !Text

Required "password"

mkFormatTest

Arguments

:: Double
formatTestNumber

-> ByteArray
formatTestByte

-> Date
formatTestDate

-> Text
formatTestPassword

-> FormatTest

Construct a value of type FormatTest (by applying it's required fields, if any)

HasOnlyReadOnly

data HasOnlyReadOnly

HasOnlyReadOnly

Constructors

HasOnlyReadOnly

Fields

hasOnlyReadOnlyBar :: !(Maybe Text)

"bar"

hasOnlyReadOnlyFoo :: !(Maybe Text)

"foo"

mkHasOnlyReadOnly :: HasOnlyReadOnly

Construct a value of type HasOnlyReadOnly (by applying it's required fields, if any)

MapTest

data MapTest

MapTest

Constructors

MapTest

Fields

mapTestMapMapOfString :: !(Maybe (Map String (Map String Text)))

"map_map_of_string"

mapTestMapOfEnumString :: !(Maybe (Map String Text))

"map_of_enum_string"

mkMapTest :: MapTest

Construct a value of type MapTest (by applying it's required fields, if any)

MixedPropertiesAndAdditionalPropertiesClass

data MixedPropertiesAndAdditionalPropertiesClass

MixedPropertiesAndAdditionalPropertiesClass

Constructors

MixedPropertiesAndAdditionalPropertiesClass

Fields

mixedPropertiesAndAdditionalPropertiesClassUuid :: !(Maybe Text)

"uuid"

mixedPropertiesAndAdditionalPropertiesClassDateTime :: !(Maybe DateTime)

"dateTime"

mixedPropertiesAndAdditionalPropertiesClassMap :: !(Maybe (Map String Animal))

"map"

mkMixedPropertiesAndAdditionalPropertiesClass :: MixedPropertiesAndAdditionalPropertiesClass

Construct a value of type MixedPropertiesAndAdditionalPropertiesClass (by applying it's required fields, if any)

Model200Response

data Model200Response

Model200Response Model for testing model name starting with number

Constructors

Model200Response

Fields

model200ResponseName :: !(Maybe Int)

"name"

model200ResponseClass :: !(Maybe Text)

"class"

mkModel200Response :: Model200Response

Construct a value of type Model200Response (by applying it's required fields, if any)

ModelList

data ModelList

ModelList

Constructors

ModelList

Fields

modelList123List :: !(Maybe Text)

"123-list"

mkModelList :: ModelList

Construct a value of type ModelList (by applying it's required fields, if any)

ModelReturn

data ModelReturn

ModelReturn Model for testing reserved words

Constructors

ModelReturn

Fields

modelReturnReturn :: !(Maybe Int)

"return"

mkModelReturn :: ModelReturn

Construct a value of type ModelReturn (by applying it's required fields, if any)

Name

data Name

Name Model for testing model name same as property name

Constructors

Name

Fields

nameName :: !Int

Required "name"

nameSnakeCase :: !(Maybe Int)

"snake_case"

nameProperty :: !(Maybe Text)

"property"

name123Number :: !(Maybe Int)

"123Number"

mkName

Arguments

:: Int
nameName

-> Name

Construct a value of type Name (by applying it's required fields, if any)

NumberOnly

data NumberOnly

NumberOnly

Constructors

NumberOnly

Fields

numberOnlyJustNumber :: !(Maybe Double)

JustNumber

mkNumberOnly :: NumberOnly

Construct a value of type NumberOnly (by applying it's required fields, if any)

Order

data Order

Order

Constructors

Order

Fields

orderId :: !(Maybe Integer)

"id"

orderPetId :: !(Maybe Integer)

"petId"

orderQuantity :: !(Maybe Int)

"quantity"

orderShipDate :: !(Maybe DateTime)

"shipDate"

orderStatus :: !(Maybe Text)

"status" - Order Status

orderComplete :: !(Maybe Bool)

"complete"

mkOrder :: Order

Construct a value of type Order (by applying it's required fields, if any)

OuterBoolean

data OuterBoolean

OuterBoolean

Constructors

OuterBoolean

mkOuterBoolean :: OuterBoolean

Construct a value of type OuterBoolean (by applying it's required fields, if any)

OuterComposite

data OuterComposite

OuterComposite

Constructors

OuterComposite

Fields

outerCompositeMyNumber :: !(Maybe OuterNumber)

"my_number"

outerCompositeMyString :: !(Maybe OuterString)

"my_string"

outerCompositeMyBoolean :: !(Maybe OuterBoolean)

"my_boolean"

mkOuterComposite :: OuterComposite

Construct a value of type OuterComposite (by applying it's required fields, if any)

OuterEnum

data OuterEnum

OuterEnum

Constructors

OuterEnum

mkOuterEnum :: OuterEnum

Construct a value of type OuterEnum (by applying it's required fields, if any)

OuterNumber

data OuterNumber

OuterNumber

Constructors

OuterNumber

mkOuterNumber :: OuterNumber

Construct a value of type OuterNumber (by applying it's required fields, if any)

OuterString

data OuterString

OuterString

Constructors

OuterString

mkOuterString :: OuterString

Construct a value of type OuterString (by applying it's required fields, if any)

Pet

data Pet

Pet

Constructors

Pet

Fields

petId :: !(Maybe Integer)

"id"

petCategory :: !(Maybe Category)

"category"

petName :: !Text

Required "name"

petPhotoUrls :: ![Text]

Required "photoUrls"

petTags :: !(Maybe [Tag])

"tags"

petStatus :: !(Maybe Text)

"status" - pet status in the store

mkPet

Arguments

:: Text
petName

-> [Text]
petPhotoUrls

-> Pet

Construct a value of type Pet (by applying it's required fields, if any)

ReadOnlyFirst

data ReadOnlyFirst

ReadOnlyFirst

Constructors

ReadOnlyFirst

Fields

readOnlyFirstBar :: !(Maybe Text)

"bar"

readOnlyFirstBaz :: !(Maybe Text)

"baz"

mkReadOnlyFirst :: ReadOnlyFirst

Construct a value of type ReadOnlyFirst (by applying it's required fields, if any)

SpecialModelName

data SpecialModelName

SpecialModelName

Constructors

SpecialModelName

Fields

specialModelNameSpecialPropertyName :: !(Maybe Integer)

"$special[property.name]"

mkSpecialModelName :: SpecialModelName

Construct a value of type SpecialModelName (by applying it's required fields, if any)

Tag

data Tag

Tag

Constructors

Tag

Fields

tagId :: !(Maybe Integer)

"id"

tagName :: !(Maybe Text)

"name"

mkTag :: Tag

Construct a value of type Tag (by applying it's required fields, if any)

User

data User

User

Constructors

User

Fields

userId :: !(Maybe Integer)

"id"

userUsername :: !(Maybe Text)

"username"

userFirstName :: !(Maybe Text)

"firstName"

userLastName :: !(Maybe Text)

"lastName"

userEmail :: !(Maybe Text)

"email"

userPassword :: !(Maybe Text)

"password"

userPhone :: !(Maybe Text)

"phone"

userUserStatus :: !(Maybe Int)

"userStatus" - User Status

mkUser :: User

Construct a value of type User (by applying it's required fields, if any)

Cat

data Cat

Cat

Constructors

Cat

Fields

catClassName :: !Text

Required "className"

catColor :: !(Maybe Text)

"color"

catDeclawed :: !(Maybe Bool)

"declawed"

mkCat

Arguments

:: Text
catClassName

-> Cat

Construct a value of type Cat (by applying it's required fields, if any)

Dog

data Dog

Dog

Constructors

Dog

Fields

dogClassName :: !Text

Required "className"

dogColor :: !(Maybe Text)

"color"

dogBreed :: !(Maybe Text)

"breed"

mkDog

Arguments

:: Text
dogClassName

-> Dog

Construct a value of type Dog (by applying it's required fields, if any)

Parameter newtypes

newtype AdditionalMetadata

Constructors

AdditionalMetadata

Fields

unAdditionalMetadata :: Text

newtype ApiKey

Constructors

ApiKey

Fields

unApiKey :: Text

newtype Body

Constructors

Body

Fields

unBody :: [User]

newtype Byte

Constructors

Byte

Fields

unByte :: ByteArray

newtype Callback

Constructors

Callback

Fields

unCallback :: Text

newtype EnumFormString

Constructors

EnumFormString

Fields

unEnumFormString :: Text

newtype EnumFormStringArray

Constructors

EnumFormStringArray

Fields

unEnumFormStringArray :: [Text]

newtype EnumHeaderString

Constructors

EnumHeaderString

Fields

unEnumHeaderString :: Text

newtype EnumHeaderStringArray

Constructors

EnumHeaderStringArray

Fields

unEnumHeaderStringArray :: [Text]

newtype EnumQueryDouble

Constructors

EnumQueryDouble

Fields

unEnumQueryDouble :: Double

newtype EnumQueryInteger

Constructors

EnumQueryInteger

Fields

unEnumQueryInteger :: Int

newtype EnumQueryString

Constructors

EnumQueryString

Fields

unEnumQueryString :: Text

newtype EnumQueryStringArray

Constructors

EnumQueryStringArray

Fields

unEnumQueryStringArray :: [Text]

newtype File

Constructors

File

Fields

unFile :: FilePath

newtype Int32

Constructors

Int32

Fields

unInt32 :: Int

newtype Int64

Constructors

Int64

Fields

unInt64 :: Integer

newtype Name2

Constructors

Name2

Fields

unName2 :: Text

newtype Number

Constructors

Number

Fields

unNumber :: Double

newtype OrderId

Constructors

OrderId

Fields

unOrderId :: Integer

newtype OrderIdText

Constructors

OrderIdText

Fields

unOrderIdText :: Text

newtype Param

Constructors

Param

Fields

unParam :: Text

newtype Param2

Constructors

Param2

Fields

unParam2 :: Text

newtype ParamBinary

Constructors

ParamBinary

Fields

unParamBinary :: Binary

newtype ParamDate

Constructors

ParamDate

Fields

unParamDate :: Date

newtype ParamDateTime

Constructors

ParamDateTime

Fields

unParamDateTime :: DateTime

newtype ParamDouble

Constructors

ParamDouble

Fields

unParamDouble :: Double

newtype ParamFloat

Constructors

ParamFloat

Fields

unParamFloat :: Float

newtype ParamInteger

Constructors

ParamInteger

Fields

unParamInteger :: Int

newtype ParamString

Constructors

ParamString

Fields

unParamString :: Text

newtype Password

Constructors

Password

Fields

unPassword :: Text

newtype PatternWithoutDelimiter

Constructors

PatternWithoutDelimiter

Fields

unPatternWithoutDelimiter :: Text

newtype PetId

Constructors

PetId

Fields

unPetId :: Integer

newtype Status

Constructors

Status

Fields

unStatus :: [Text]

newtype StatusText

Constructors

StatusText

Fields

unStatusText :: Text

newtype Tags

Constructors

Tags

Fields

unTags :: [Text]

newtype Username

Constructors

Username

Fields

unUsername :: Text

Utils

_omitNulls :: [(Text, Value)] -> Value

Removes Null fields. (OpenAPI-Specification 2.0 does not allow Null in JSON)

_toFormItem :: (ToHttpApiData a, Functor f) => t -> f a -> f (t, [Text])

Encodes fields using WH.toQueryParam

_emptyToNothing :: Maybe String -> Maybe String

Collapse (Just "") to Nothing

_memptyToNothing :: (Monoid a, Eq a) => Maybe a -> Maybe a

Collapse (Just mempty) to Nothing

DateTime Formatting

newtype DateTime

Constructors

DateTime

Fields

unDateTime :: UTCTime

_readDateTime :: (ParseTime t, Monad m, Alternative m) => String -> m t

_parseISO8601

_showDateTime :: (t ~ UTCTime, FormatTime t) => t -> String

TI.formatISO8601Millis

_parseISO8601 :: (ParseTime t, Monad m, Alternative m) => String -> m t

parse an ISO8601 date-time string

Date Formatting

newtype Date

Constructors

Date

Fields

unDate :: Day

_readDate :: (ParseTime t, Monad m) => String -> m t

TI.parseTimeM True TI.defaultTimeLocale "%Y-%m-%d"

_showDate :: FormatTime t => t -> String

TI.formatTime TI.defaultTimeLocale "%Y-%m-%d"

Byte/Binary Formatting

newtype ByteArray

base64 encoded characters

Constructors

ByteArray

Fields

unByteArray :: ByteString

_readByteArray :: Monad m => Text -> m ByteArray

read base64 encoded characters

_showByteArray :: ByteArray -> Text

show base64 encoded characters

newtype Binary

any sequence of octets

Constructors

Binary

Fields

unBinary :: ByteString

_readBinaryBase64 :: Monad m => Text -> m Binary

_showBinaryBase64 :: Binary -> Text