SwaggerPetstore.API

Description

Operations

addPet

addPet

Arguments

:: UseContentType AddPet Pet contenttype

=> contenttype
request content-type (mimetype)

-> Pet
"body" - Pet object that needs to be added to the store

-> SwaggerPetstoreRequest AddPet contenttype ()

POST /pet

Add a new pet to the store

AuthMethod: petstore_auth

data AddPet

updatePetWithForm

updatePetWithForm

Arguments

:: Integer
"petId" - ID of pet that needs to be updated

-> SwaggerPetstoreRequest UpdatePetWithForm MimeFormUrlEncoded ()

POST /pet/{petId}

Updates a pet in the store with form data

AuthMethod: petstore_auth

data UpdatePetWithForm

HasOptionalParam

class HasOptionalParam req param where

Designates the optional parameters of a request

Minimal complete definition

applyOptionalParam | (-&-)

Methods

applyOptionalParam :: SwaggerPetstoreRequest req i res -> param -> SwaggerPetstoreRequest req i res

Apply an optional parameter to a request

(-&-) :: SwaggerPetstoreRequest req i res -> param -> SwaggerPetstoreRequest req i res infixl 2

infix operator / alias for addOptionalParam

class HasOptionalBodyParam req param where

Minimal complete definition

setBodyParam

Methods

setBodyParam :: UseContentType req param i => SwaggerPetstoreRequest req i res -> param -> SwaggerPetstoreRequest req i res

Optional Request Parameter Types

newtype Name

Constructors

Name

Fields

unName :: Text

newtype Status

Constructors

Status

Fields

unStatus :: Text

SwaggerPetstoreRequest

data SwaggerPetstoreRequest req i res

Represents a request. The "req" type variable is the request type. The "res" type variable is the response type.

Constructors

SwaggerPetstoreRequest

Fields

rMethod :: Method

Method of SwaggerPetstoreRequest

urlPath :: [ByteString]

Endpoint of SwaggerPetstoreRequest

params :: Params

params of SwaggerPetstoreRequest

ctype :: i

content-type of SwaggerPetstoreRequest

data Params

Request Params

Constructors

Params

Fields

paramsQuery :: Query

paramsHeaders :: RequestHeaders

paramsBody :: ParamBody

data ParamBody

Request Body

Constructors

ParamBodyNone

ParamBodyB ByteString

ParamBodyBL ByteString

ParamBodyFormUrlEncoded Form

ParamBodyMultipartFormData [Part]

SwaggerPetstoreRequest Utils

_mkRequest

Arguments

:: Method
Method

-> [ByteString]
Endpoint

-> ctype
Content-Type

-> SwaggerPetstoreRequest req ctype res
req: Request Type, res: Response Type

_mkParams :: Params

_addHeader :: SwaggerPetstoreRequest req i res -> [Header] -> SwaggerPetstoreRequest req i res

_addQuery :: SwaggerPetstoreRequest req i res -> [QueryItem] -> SwaggerPetstoreRequest req i res

_addForm :: SwaggerPetstoreRequest req i res -> Form -> SwaggerPetstoreRequest req i res

_addMultiFormPart :: SwaggerPetstoreRequest req MimeMultipartFormData res -> Part -> SwaggerPetstoreRequest req MimeMultipartFormData res

_setBodyBS :: SwaggerPetstoreRequest req i res -> ByteString -> SwaggerPetstoreRequest req i res

_setBodyLBS :: SwaggerPetstoreRequest req i res -> ByteString -> SwaggerPetstoreRequest req i res

Params Utils

toPath :: ToHttpApiData a => a -> ByteString

toHeader :: ToHttpApiData a => (HeaderName, a) -> [Header]

toForm :: ToHttpApiData v => (ByteString, v) -> Form

toQuery :: ToHttpApiData a => (ByteString, Maybe a) -> [QueryItem]

Swagger CollectionFormat Utils

data CollectionFormat

Determines the format of the array if type array is used.

Constructors

CommaSeparated
CSV format for multiple parameters.

SpaceSeparated
Also called SSV

TabSeparated
Also called TSV

PipeSeparated
`value1|value2|value2`

MultiParamArray
Using multiple GET parameters, e.g. `foo=bar&foo=baz`. This is valid only for parameters in "query" (Query) or "formData" (Form)

toHeaderColl :: ToHttpApiData a => CollectionFormat -> (HeaderName, [a]) -> [Header]

toFormColl :: ToHttpApiData v => CollectionFormat -> (ByteString, [v]) -> Form

toQueryColl :: ToHttpApiData a => CollectionFormat -> (ByteString, Maybe [a]) -> Query

_toColl :: Traversable f => CollectionFormat -> (f a -> [(b, ByteString)]) -> f [a] -> [(b, ByteString)]

_toCollA :: (Traversable f, Traversable t, Alternative t) => CollectionFormat -> (f (t a) -> [(b, t ByteString)]) -> f (t [a]) -> [(b, t ByteString)]

_toCollA' :: (Monoid c, Traversable f, Traversable t, Alternative t) => CollectionFormat -> (f (t a) -> [(b, t c)]) -> (Char -> c) -> f (t [a]) -> [(b, t c)]

Content Negotiation

UseContentType constraint

type UseContentType operation model mediatype = (Consumes operation mediatype, MimeRender mediatype model)

Combines Consumes & MimeRender for an operation

Mime Types

data MimeJSON

Constructors

MimeJSON

data MimeXML

Constructors

MimeXML

data MimePlainText

Constructors

MimePlainText

data MimeFormUrlEncoded

Constructors

MimeFormUrlEncoded

data MimeMultipartFormData

Constructors

MimeMultipartFormData

data MimeNoContent

Constructors

MimeNoContent

MimeType Class

class MimeType mtype where

Minimal complete definition

mimeType | mimeTypes

Methods

mimeType :: mtype -> Maybe MediaType

mimeTypes :: mtype -> [MediaType]

MimeType Instances

MimeRender Class

class MimeType mtype => MimeRender mtype i where

Minimal complete definition

mimeRender

Methods

mimeRender :: mtype -> i -> ByteString

MimeRender Instances

MimeUnrender Class

class MimeType mtype => MimeUnrender mtype o where

Minimal complete definition

mimeUnrender

Methods

mimeUnrender :: mtype -> ByteString -> Either String o

MimeUnrender Instances

Request Consumes

class Consumes req mtype

Request Produces

class Produces req mtype