SwaggerPetstore.API

Description

Operations

addPet

addPet

Arguments

:: Pet
"body" - Pet object that needs to be added to the store

-> SwaggerPetstoreRequest AddPet ()

POST /pet

Add a new pet to the store

AuthMethod: petstore_auth

Consumes: applicationjson, applicationxml

Produces: applicationxml, applicationjson

data AddPet

deletePet

deletePet

Arguments

:: Integer
"petId" - Pet id to delete

-> SwaggerPetstoreRequest DeletePet ()

DELETE /pet/{petId}

Deletes a pet

AuthMethod: petstore_auth

Produces: applicationxml, applicationjson

Optional Params:

"api_key" :: Text

data DeletePet

findPetsByStatus

findPetsByStatus

Arguments

:: [Text]
"status" - Status values that need to be considered for filter

-> SwaggerPetstoreRequest FindPetsByStatus [Pet]

GET /pet/findByStatus

Finds Pets by status

Multiple status values can be provided with comma separated strings

AuthMethod: petstore_auth

Produces: applicationxml, applicationjson

data FindPetsByStatus

findPetsByTags

findPetsByTags

Arguments

:: [Text]
"tags" - Tags to filter by

-> SwaggerPetstoreRequest FindPetsByTags [Pet]

Deprecated:

GET /pet/findByTags

Finds Pets by tags

Muliple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.

AuthMethod: petstore_auth

Produces: applicationxml, applicationjson

data FindPetsByTags

getPetById

getPetById

Arguments

:: Integer
"petId" - ID of pet to return

-> SwaggerPetstoreRequest GetPetById Pet

GET /pet/{petId}

Find pet by ID

Returns a single pet

AuthMethod: api_key

Produces: applicationxml, applicationjson

data GetPetById

updatePet

updatePet

Arguments

:: Pet
"body" - Pet object that needs to be added to the store

-> SwaggerPetstoreRequest UpdatePet ()

PUT /pet

Update an existing pet

AuthMethod: petstore_auth

Consumes: applicationjson, applicationxml

Produces: applicationxml, applicationjson

data UpdatePet

updatePetWithForm

updatePetWithForm

Arguments

:: Integer
"petId" - ID of pet that needs to be updated

-> SwaggerPetstoreRequest UpdatePetWithForm ()

POST /pet/{petId}

Updates a pet in the store with form data

AuthMethod: petstore_auth

Consumes: application/x-www-form-urlencoded

Produces: applicationxml, applicationjson

Optional Params:

"name" :: Text - Updated name of the pet

"status" :: Text - Updated status of the pet

data UpdatePetWithForm

uploadFile

uploadFile

Arguments

:: Integer
"petId" - ID of pet to update

-> SwaggerPetstoreRequest UploadFile ApiResponse

POST /pet/{petId}/uploadImage

uploads an image

AuthMethod: petstore_auth

Consumes: multipart/form-data

Produces: application/json

Optional Params:

"additionalMetadata" :: Text - Additional data to pass to server

"file" :: FilePath - file to upload

data UploadFile

deleteOrder

deleteOrder

Arguments

:: Integer
"orderId" - ID of the order that needs to be deleted

-> SwaggerPetstoreRequest DeleteOrder ()

DELETE /store/order/{orderId}

Delete purchase order by ID

For valid response try integer IDs with positive integer value. Negative or non-integer values will generate API errors

Produces: applicationxml, applicationjson

data DeleteOrder

getInventory

getInventory :: SwaggerPetstoreRequest GetInventory (Map String Int)

GET /store/inventory

Returns pet inventories by status

Returns a map of status codes to quantities

AuthMethod: api_key

Produces: application/json

data GetInventory

getOrderById

getOrderById

Arguments

:: Integer
"orderId" - ID of pet that needs to be fetched

-> SwaggerPetstoreRequest GetOrderById Order

GET /store/order/{orderId}

Find purchase order by ID

For valid response try integer IDs with value >= 1 and <= 10. Other values will generated exceptions

Produces: applicationxml, applicationjson

data GetOrderById

placeOrder

placeOrder

Arguments

:: Order
"body" - order placed for purchasing the pet

-> SwaggerPetstoreRequest PlaceOrder Order

POST /store/order

Place an order for a pet

Produces: applicationxml, applicationjson

data PlaceOrder

createUser

createUser

Arguments

:: User
"body" - Created user object

-> SwaggerPetstoreRequest CreateUser ()

POST /user

Create user

This can only be done by the logged in user.

Produces: applicationxml, applicationjson

data CreateUser

createUsersWithArrayInput

createUsersWithArrayInput

Arguments

:: [User]
"body" - List of user object

-> SwaggerPetstoreRequest CreateUsersWithArrayInput ()

POST /user/createWithArray

Creates list of users with given input array

Produces: applicationxml, applicationjson

data CreateUsersWithArrayInput

createUsersWithListInput

createUsersWithListInput

Arguments

:: [User]
"body" - List of user object

-> SwaggerPetstoreRequest CreateUsersWithListInput ()

POST /user/createWithList

Creates list of users with given input array

Produces: applicationxml, applicationjson

data CreateUsersWithListInput

deleteUser

deleteUser

Arguments

:: Text
"username" - The name that needs to be deleted

-> SwaggerPetstoreRequest DeleteUser ()

DELETE /user/{username}

Delete user

This can only be done by the logged in user.

Produces: applicationxml, applicationjson

data DeleteUser

getUserByName

getUserByName

Arguments

:: Text
"username" - The name that needs to be fetched. Use user1 for testing.

-> SwaggerPetstoreRequest GetUserByName User

GET /user/{username}

Get user by user name

Produces: applicationxml, applicationjson

data GetUserByName

loginUser

loginUser

Arguments

:: Text
"username" - The user name for login

-> Text
"password" - The password for login in clear text

-> SwaggerPetstoreRequest LoginUser Text

GET /user/login

Logs user into the system

Produces: applicationxml, applicationjson

data LoginUser

logoutUser

logoutUser :: SwaggerPetstoreRequest LogoutUser ()

GET /user/logout

Logs out current logged in user session

Produces: applicationxml, applicationjson

data LogoutUser

updateUser

updateUser

Arguments

:: Text
"username" - name that need to be updated

-> User
"body" - Updated user object

-> SwaggerPetstoreRequest UpdateUser ()

PUT /user/{username}

Updated user

This can only be done by the logged in user.

Produces: applicationxml, applicationjson

data UpdateUser

SwaggerPetstoreRequest

data SwaggerPetstoreRequest req res

Constructors

SwaggerPetstoreRequest

Fields

rMethod :: Method

Method of SwaggerPetstoreRequest

endpoint :: [Text]

Endpoint of SwaggerPetstoreRequest

params :: [EncodedParam]

Encoded params of SwaggerPetstoreRequest

mkSwaggerPetstoreRequest :: Method -> [Text] -> [EncodedParam] -> SwaggerPetstoreRequest req res

type TupleBS8 = (ByteString, ByteString)

Type alias for query parameters

Params

HasOptionalParam

class ToEncodedParam param => HasOptionalParam request param

addOptionalParam

addOptionalParam :: HasOptionalParam req param => SwaggerPetstoreRequest req res -> param -> SwaggerPetstoreRequest req res

Add an optional parameter to a request

(-&-)

(-&-) :: HasOptionalParam req param => SwaggerPetstoreRequest req res -> param -> SwaggerPetstoreRequest req res

infix operator for addOptionalParam

ToEncodedParam

class ToEncodedParam param where

Minimal complete definition

toEncodedParam

Methods

toEncodedParam :: param -> [EncodedParam] -> [EncodedParam]

EncodedParam

data EncodedParam

Constructors

Query TupleBS8

Body ByteString

Header TupleBS8

data ResultFormatType

Constructors

FormatJson

FormatXml

toPath :: IsString t => t1 -> t