SwaggerPetstore.API

Description

Operations

addPet

addPet

Arguments

:: Pet
"body" - Pet object that needs to be added to the store

-> SwaggerPetstoreRequest AddPet ()

POST /pet

Add a new pet to the store

AuthMethod: petstore_auth

Consumes: applicationjson, applicationxml

Produces: applicationxml, applicationjson

data AddPet

deletePet

deletePet

Arguments

:: Integer
"petId" - Pet id to delete

-> SwaggerPetstoreRequest DeletePet ()

DELETE /pet/{petId}

Deletes a pet

AuthMethod: petstore_auth

Produces: applicationxml, applicationjson

data DeletePet

findPetsByStatus

findPetsByStatus

Arguments

:: [Text]
"status" - Status values that need to be considered for filter

-> SwaggerPetstoreRequest FindPetsByStatus [Pet]

GET /pet/findByStatus

Finds Pets by status

Multiple status values can be provided with comma separated strings

AuthMethod: petstore_auth

Produces: applicationxml, applicationjson

data FindPetsByStatus

findPetsByTags

findPetsByTags

Arguments

:: [Text]
"tags" - Tags to filter by

-> SwaggerPetstoreRequest FindPetsByTags [Pet]

Deprecated:

GET /pet/findByTags

Finds Pets by tags

Muliple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.

AuthMethod: petstore_auth

Produces: applicationxml, applicationjson

data FindPetsByTags

getPetById

getPetById

Arguments

:: Integer
"petId" - ID of pet to return

-> SwaggerPetstoreRequest GetPetById Pet

GET /pet/{petId}

Find pet by ID

Returns a single pet

AuthMethod: api_key

Produces: applicationxml, applicationjson

data GetPetById

updatePet

updatePet

Arguments

:: Pet
"body" - Pet object that needs to be added to the store

-> SwaggerPetstoreRequest UpdatePet ()

PUT /pet

Update an existing pet

AuthMethod: petstore_auth

Consumes: applicationjson, applicationxml

Produces: applicationxml, applicationjson

data UpdatePet

updatePetWithForm

updatePetWithForm

Arguments

:: Integer
"petId" - ID of pet that needs to be updated

-> SwaggerPetstoreRequest UpdatePetWithForm ()

POST /pet/{petId}

Updates a pet in the store with form data

AuthMethod: petstore_auth

Consumes: application/x-www-form-urlencoded

Produces: applicationxml, applicationjson

data UpdatePetWithForm

uploadFile

uploadFile

Arguments

:: Integer
"petId" - ID of pet to update

-> SwaggerPetstoreRequest UploadFile ApiResponse

POST /pet/{petId}/uploadImage

uploads an image

AuthMethod: petstore_auth

Consumes: multipart/form-data

Produces: application/json

data UploadFile

deleteOrder

deleteOrder

Arguments

:: Integer
"orderId" - ID of the order that needs to be deleted

-> SwaggerPetstoreRequest DeleteOrder ()

DELETE /store/order/{orderId}

Delete purchase order by ID

For valid response try integer IDs with positive integer value. Negative or non-integer values will generate API errors

Produces: applicationxml, applicationjson

data DeleteOrder

getInventory

getInventory :: SwaggerPetstoreRequest GetInventory (Map String Int)

GET /store/inventory

Returns pet inventories by status

Returns a map of status codes to quantities

AuthMethod: api_key

Produces: application/json

data GetInventory

getOrderById

getOrderById

Arguments

:: Integer
"orderId" - ID of pet that needs to be fetched

-> SwaggerPetstoreRequest GetOrderById Order

GET /store/order/{orderId}

Find purchase order by ID

For valid response try integer IDs with value >= 1 and <= 10. Other values will generated exceptions

Produces: applicationxml, applicationjson

data GetOrderById

placeOrder

placeOrder

Arguments

:: Order
"body" - order placed for purchasing the pet

-> SwaggerPetstoreRequest PlaceOrder Order

POST /store/order

Place an order for a pet

Produces: applicationxml, applicationjson

data PlaceOrder

createUser

createUser

Arguments

:: User
"body" - Created user object

-> SwaggerPetstoreRequest CreateUser ()

POST /user

Create user

This can only be done by the logged in user.

Produces: applicationxml, applicationjson

data CreateUser

createUsersWithArrayInput

createUsersWithArrayInput

Arguments

:: [User]
"body" - List of user object

-> SwaggerPetstoreRequest CreateUsersWithArrayInput ()

POST /user/createWithArray

Creates list of users with given input array

Produces: applicationxml, applicationjson

data CreateUsersWithArrayInput

createUsersWithListInput

createUsersWithListInput

Arguments

:: [User]
"body" - List of user object

-> SwaggerPetstoreRequest CreateUsersWithListInput ()

POST /user/createWithList

Creates list of users with given input array

Produces: applicationxml, applicationjson

data CreateUsersWithListInput

deleteUser

deleteUser

Arguments

:: Text
"username" - The name that needs to be deleted

-> SwaggerPetstoreRequest DeleteUser ()

DELETE /user/{username}

Delete user

This can only be done by the logged in user.

Produces: applicationxml, applicationjson

data DeleteUser

getUserByName

getUserByName

Arguments

:: Text
"username" - The name that needs to be fetched. Use user1 for testing.

-> SwaggerPetstoreRequest GetUserByName User

GET /user/{username}

Get user by user name

Produces: applicationxml, applicationjson

data GetUserByName

loginUser

loginUser

Arguments

:: Text
"username" - The user name for login

-> Text
"password" - The password for login in clear text

-> SwaggerPetstoreRequest LoginUser Text

GET /user/login

Logs user into the system

Produces: applicationxml, applicationjson

data LoginUser

logoutUser

logoutUser :: SwaggerPetstoreRequest LogoutUser ()

GET /user/logout

Logs out current logged in user session

Produces: applicationxml, applicationjson

data LogoutUser

updateUser

updateUser

Arguments

:: Text
"username" - name that need to be updated

-> User
"body" - Updated user object

-> SwaggerPetstoreRequest UpdateUser ()

PUT /user/{username}

Updated user

This can only be done by the logged in user.

Produces: applicationxml, applicationjson

data UpdateUser

HasOptionalParam

class HasOptionalParam req param where

Designates the optional parameters of a request

Minimal complete definition

applyOptionalParam | (-&-)

Methods

applyOptionalParam :: forall res. SwaggerPetstoreRequest req res -> param -> SwaggerPetstoreRequest req res

Apply an optional parameter to a request

(-&-) :: forall res. SwaggerPetstoreRequest req res -> param -> SwaggerPetstoreRequest req res infixl 2

infix operator / alias for addOptionalParam

Optional Request Parameter Types

newtype Api'Underscorekey

Constructors

Api'Underscorekey

Fields

unApi'Underscorekey :: Text

newtype Name

Constructors

Name

Fields

unName :: Text

newtype Status

Constructors

Status

Fields

unStatus :: Text

newtype AdditionalMetadata

Constructors

AdditionalMetadata

Fields

unAdditionalMetadata :: Text

newtype File

Constructors

File

Fields

unFile :: FilePath

SwaggerPetstoreRequest

data SwaggerPetstoreRequest req res

Represents a request. The "req" type variable is the request type. The "res" type variable is the response type.

Constructors

SwaggerPetstoreRequest

Fields

rMethod :: Method

Method of SwaggerPetstoreRequest

urlPath :: [ByteString]

Endpoint of SwaggerPetstoreRequest

params :: Params

params of SwaggerPetstoreRequest

data Params

Request Params

Constructors

Params

Fields

paramsQuery :: Query

paramsHeaders :: RequestHeaders

paramsBody :: ParamBody

data ParamBody

Request Body

Constructors

ParamBodyNone

ParamBodyForm Form

ParamBodyMultiForm [Part]

ParamBodyB ByteString

ParamBodyBL ByteString

SwaggerPetstoreRequest Utils

_mkRequest

Arguments

:: Method
Method

-> [ByteString]
Endpoint

-> SwaggerPetstoreRequest req res
req: Request Type, res: Response Type

_mkParams :: Params

_addHeader :: SwaggerPetstoreRequest req res -> [Header] -> SwaggerPetstoreRequest req res

_addQuery :: SwaggerPetstoreRequest req res -> [QueryItem] -> SwaggerPetstoreRequest req res

_addForm :: SwaggerPetstoreRequest req res -> Form -> SwaggerPetstoreRequest req res

_addMultiFormPart :: SwaggerPetstoreRequest req res -> Part -> SwaggerPetstoreRequest req res

_setBodyBS :: SwaggerPetstoreRequest req res -> ByteString -> SwaggerPetstoreRequest req res

_setBodyLBS :: SwaggerPetstoreRequest req res -> ByteString -> SwaggerPetstoreRequest req res

Params Utils

toPath :: ToHttpApiData a => a -> ByteString

toHeader :: ToHttpApiData a => (HeaderName, a) -> [Header]

toForm :: ToHttpApiData v => (ByteString, v) -> Form

toQuery :: ToHttpApiData a => (ByteString, Maybe a) -> [QueryItem]

Swagger CollectionFormat Utils

data CollectionFormat

Determines the format of the array if type array is used.

Constructors

CommaSeparated
CSV format for multiple parameters.

SpaceSeparated
Also called SSV

TabSeparated
Also called TSV

PipeSeparated
`value1|value2|value2`

MultiParamArray
Using multiple GET parameters, e.g. `foo=bar&foo=baz`. This is valid only for parameters in "query" (Query) or "formData" (Form)

toHeaderColl :: ToHttpApiData a => CollectionFormat -> (HeaderName, [a]) -> [Header]

toFormColl :: ToHttpApiData v => CollectionFormat -> (ByteString, [v]) -> Form

toQueryColl :: ToHttpApiData a => CollectionFormat -> (ByteString, Maybe [a]) -> Query

_toColl :: Traversable f => CollectionFormat -> (f a -> [(b, ByteString)]) -> f [a] -> [(b, ByteString)]

_toCollA :: (Traversable f, Traversable t, Alternative t) => CollectionFormat -> (f (t a) -> [(b, t ByteString)]) -> f (t [a]) -> [(b, t ByteString)]

_toCollA' :: (Monoid c, Traversable f, Traversable t, Alternative t) => CollectionFormat -> (f (t a) -> [(b, t c)]) -> (Char -> c) -> f (t [a]) -> [(b, t c)]

Mime Types

data JSON

data XML

data PlainText

data FormUrlEncoded

data NoContent

class MimeType mtype where

Minimal complete definition

mimeType | mimeTypes

Methods

mimeType :: Proxy mtype -> MediaType

mimeTypes :: Proxy mtype -> [MediaType]

MimeRender

class MimeType mtype => MimeRender mtype a where

Minimal complete definition

mimeRender

Methods

mimeRender :: Proxy mtype -> a -> ByteString

MimeUnrender

class MimeType mtype => MimeUnrender mtype a where

Minimal complete definition

mimeUnrender | mimeUnrenderWithType

Methods

mimeUnrender :: Proxy mtype -> ByteString -> Either String a

mimeUnrenderWithType :: Proxy mtype -> MediaType -> ByteString -> Either String a

Consumes

class Consumes req mtype

Produces

class Produces req mtype