SwaggerPetstore.API

Description

Operations

AnotherFake

testSpecialTags

testSpecialTags

Arguments

:: (Consumes TestSpecialTags contentType, MimeRender contentType Client)

=> contentType
request content-type (MimeType)

-> Client
"body" - client model

-> SwaggerPetstoreRequest TestSpecialTags contentType Client

PATCH /another-fake/dummy

To test special tags

To test special tags

data TestSpecialTags

Fake

fakeOuterBooleanSerialize

fakeOuterBooleanSerialize

Arguments

:: Consumes FakeOuterBooleanSerialize contentType

=> contentType
request content-type (MimeType)

-> SwaggerPetstoreRequest FakeOuterBooleanSerialize contentType OuterBoolean

POST /fake/outer/boolean

Test serialization of outer boolean types

data FakeOuterBooleanSerialize

fakeOuterCompositeSerialize

fakeOuterCompositeSerialize

Arguments

:: Consumes FakeOuterCompositeSerialize contentType

=> contentType
request content-type (MimeType)

-> SwaggerPetstoreRequest FakeOuterCompositeSerialize contentType OuterComposite

POST /fake/outer/composite

Test serialization of object with outer number type

data FakeOuterCompositeSerialize

fakeOuterNumberSerialize

fakeOuterNumberSerialize

Arguments

:: Consumes FakeOuterNumberSerialize contentType

=> contentType
request content-type (MimeType)

-> SwaggerPetstoreRequest FakeOuterNumberSerialize contentType OuterNumber

POST /fake/outer/number

Test serialization of outer number types

data FakeOuterNumberSerialize

fakeOuterStringSerialize

fakeOuterStringSerialize

Arguments

:: Consumes FakeOuterStringSerialize contentType

=> contentType
request content-type (MimeType)

-> SwaggerPetstoreRequest FakeOuterStringSerialize contentType OuterString

POST /fake/outer/string

Test serialization of outer string types

data FakeOuterStringSerialize

testClientModel

testClientModel

Arguments

:: (Consumes TestClientModel contentType, MimeRender contentType Client)

=> contentType
request content-type (MimeType)

-> Client
"body" - client model

-> SwaggerPetstoreRequest TestClientModel contentType Client

PATCH /fake

To test "client" model

To test "client" model

data TestClientModel

testEndpointParameters

testEndpointParameters

Arguments

:: Consumes TestEndpointParameters contentType

=> contentType
request content-type (MimeType)

-> Number
"number" - None

-> ParamDouble
"double" - None

-> PatternWithoutDelimiter
"patternWithoutDelimiter" - None

-> Byte
"byte" - None

-> SwaggerPetstoreRequest TestEndpointParameters contentType res

POST /fake

Fake endpoint for testing various parameters 假端點 偽のエンドポイント 가짜 엔드 포인트

Fake endpoint for testing various parameters 假端點 偽のエンドポイント 가짜 엔드 포인트

AuthMethod: AuthBasicHttpBasicTest

Note: Has Produces instances, but no response schema

data TestEndpointParameters

testEnumParameters

testEnumParameters

Arguments

:: Consumes TestEnumParameters contentType

=> contentType
request content-type (MimeType)

-> SwaggerPetstoreRequest TestEnumParameters contentType res

GET /fake

To test enum parameters

To test enum parameters

Note: Has Produces instances, but no response schema

data TestEnumParameters

testJsonFormData

testJsonFormData

Arguments

:: Consumes TestJsonFormData contentType

=> contentType
request content-type (MimeType)

-> Param
"param" - field1

-> Param2
"param2" - field2

-> SwaggerPetstoreRequest TestJsonFormData contentType NoContent

GET /fake/jsonFormData

test json serialization of form data

data TestJsonFormData

FakeClassnameTags123

testClassname

testClassname

Arguments

:: (Consumes TestClassname contentType, MimeRender contentType Client)

=> contentType
request content-type (MimeType)

-> Client
"body" - client model

-> SwaggerPetstoreRequest TestClassname contentType Client

PATCH /fake_classname_test

To test class name in snake case

AuthMethod: AuthApiKeyApiKeyQuery

data TestClassname

Pet

addPet

addPet

Arguments

:: (Consumes AddPet contentType, MimeRender contentType Pet)

=> contentType
request content-type (MimeType)

-> Pet
"body" - Pet object that needs to be added to the store

-> SwaggerPetstoreRequest AddPet contentType res

POST /pet

Add a new pet to the store

AuthMethod: AuthOAuthPetstoreAuth

Note: Has Produces instances, but no response schema

data AddPet

deletePet

deletePet

Arguments

:: PetId
"petId" - Pet id to delete

-> SwaggerPetstoreRequest DeletePet MimeNoContent res

DELETE /pet/{petId}

Deletes a pet

AuthMethod: AuthOAuthPetstoreAuth

Note: Has Produces instances, but no response schema

data DeletePet

findPetsByStatus

findPetsByStatus

Arguments

:: Status
"status" - Status values that need to be considered for filter

-> SwaggerPetstoreRequest FindPetsByStatus MimeNoContent [Pet]

GET /pet/findByStatus

Finds Pets by status

Multiple status values can be provided with comma separated strings

AuthMethod: AuthOAuthPetstoreAuth

data FindPetsByStatus

findPetsByTags

findPetsByTags

Arguments

:: Tags
"tags" - Tags to filter by

-> SwaggerPetstoreRequest FindPetsByTags MimeNoContent [Pet]

Deprecated:

GET /pet/findByTags

Finds Pets by tags

Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.

AuthMethod: AuthOAuthPetstoreAuth

data FindPetsByTags

getPetById

getPetById

Arguments

:: PetId
"petId" - ID of pet to return

-> SwaggerPetstoreRequest GetPetById MimeNoContent Pet

GET /pet/{petId}

Find pet by ID

Returns a single pet

AuthMethod: AuthApiKeyApiKey

data GetPetById

updatePet

updatePet

Arguments

:: (Consumes UpdatePet contentType, MimeRender contentType Pet)

=> contentType
request content-type (MimeType)

-> Pet
"body" - Pet object that needs to be added to the store

-> SwaggerPetstoreRequest UpdatePet contentType res

PUT /pet

Update an existing pet

AuthMethod: AuthOAuthPetstoreAuth

Note: Has Produces instances, but no response schema

data UpdatePet

updatePetWithForm

updatePetWithForm

Arguments

:: Consumes UpdatePetWithForm contentType

=> contentType
request content-type (MimeType)

-> PetId
"petId" - ID of pet that needs to be updated

-> SwaggerPetstoreRequest UpdatePetWithForm contentType res

POST /pet/{petId}

Updates a pet in the store with form data

AuthMethod: AuthOAuthPetstoreAuth

Note: Has Produces instances, but no response schema

data UpdatePetWithForm

uploadFile

uploadFile

Arguments

:: Consumes UploadFile contentType

=> contentType
request content-type (MimeType)

-> PetId
"petId" - ID of pet to update

-> SwaggerPetstoreRequest UploadFile contentType ApiResponse

POST /pet/{petId}/uploadImage

uploads an image

AuthMethod: AuthOAuthPetstoreAuth

data UploadFile

Store

deleteOrder

deleteOrder

Arguments

:: OrderIdText
"orderId" - ID of the order that needs to be deleted

-> SwaggerPetstoreRequest DeleteOrder MimeNoContent res

DELETE /store/order/{order_id}

Delete purchase order by ID

For valid response try integer IDs with value < 1000. Anything above 1000 or nonintegers will generate API errors

Note: Has Produces instances, but no response schema

data DeleteOrder

getInventory

getInventory :: SwaggerPetstoreRequest GetInventory MimeNoContent (Map String Int)

GET /store/inventory

Returns pet inventories by status

Returns a map of status codes to quantities

AuthMethod: AuthApiKeyApiKey

data GetInventory

getOrderById

getOrderById

Arguments

:: OrderId
"orderId" - ID of pet that needs to be fetched

-> SwaggerPetstoreRequest GetOrderById MimeNoContent Order

GET /store/order/{order_id}

Find purchase order by ID

For valid response try integer IDs with value 5 or 10. Other values will generated exceptions

data GetOrderById

placeOrder

placeOrder

Arguments

:: (Consumes PlaceOrder contentType, MimeRender contentType Order)

=> contentType
request content-type (MimeType)

-> Order
"body" - order placed for purchasing the pet

-> SwaggerPetstoreRequest PlaceOrder contentType Order

POST /store/order

Place an order for a pet

data PlaceOrder

User

createUser

createUser

Arguments

:: (Consumes CreateUser contentType, MimeRender contentType User)

=> contentType
request content-type (MimeType)

-> User
"body" - Created user object

-> SwaggerPetstoreRequest CreateUser contentType res

POST /user

Create user

This can only be done by the logged in user.

Note: Has Produces instances, but no response schema

data CreateUser

createUsersWithArrayInput

createUsersWithArrayInput

Arguments

:: (Consumes CreateUsersWithArrayInput contentType, MimeRender contentType Body)

=> contentType
request content-type (MimeType)

-> Body
"body" - List of user object

-> SwaggerPetstoreRequest CreateUsersWithArrayInput contentType res

POST /user/createWithArray

Creates list of users with given input array

Note: Has Produces instances, but no response schema

data CreateUsersWithArrayInput

createUsersWithListInput

createUsersWithListInput

Arguments

:: (Consumes CreateUsersWithListInput contentType, MimeRender contentType Body)

=> contentType
request content-type (MimeType)

-> Body
"body" - List of user object

-> SwaggerPetstoreRequest CreateUsersWithListInput contentType res

POST /user/createWithList

Creates list of users with given input array

Note: Has Produces instances, but no response schema

data CreateUsersWithListInput

deleteUser

deleteUser

Arguments

:: Username
"username" - The name that needs to be deleted

-> SwaggerPetstoreRequest DeleteUser MimeNoContent res

DELETE /user/{username}

Delete user

This can only be done by the logged in user.

Note: Has Produces instances, but no response schema

data DeleteUser

getUserByName

getUserByName

Arguments

:: Username
"username" - The name that needs to be fetched. Use user1 for testing.

-> SwaggerPetstoreRequest GetUserByName MimeNoContent User

GET /user/{username}

Get user by user name

data GetUserByName

loginUser

loginUser

Arguments

:: Username
"username" - The user name for login

-> Password
"password" - The password for login in clear text

-> SwaggerPetstoreRequest LoginUser MimeNoContent Text

GET /user/login

Logs user into the system

data LoginUser

logoutUser

logoutUser :: SwaggerPetstoreRequest LogoutUser MimeNoContent res

GET /user/logout

Logs out current logged in user session

Note: Has Produces instances, but no response schema

data LogoutUser

updateUser

updateUser

Arguments

:: (Consumes UpdateUser contentType, MimeRender contentType User)

=> contentType
request content-type (MimeType)

-> Username
"username" - name that need to be deleted

-> User
"body" - Updated user object

-> SwaggerPetstoreRequest UpdateUser contentType res

PUT /user/{username}

Updated user

This can only be done by the logged in user.

Note: Has Produces instances, but no response schema

data UpdateUser

HasBodyParam

class HasBodyParam req param where

Designates the body parameter of a request

Methods

setBodyParam :: forall contentType res. (Consumes req contentType, MimeRender contentType param) => SwaggerPetstoreRequest req contentType res -> param -> SwaggerPetstoreRequest req contentType res

HasOptionalParam

class HasOptionalParam req param where

Designates the optional parameters of a request

Minimal complete definition

applyOptionalParam | (-&-)

Methods

applyOptionalParam :: SwaggerPetstoreRequest req contentType res -> param -> SwaggerPetstoreRequest req contentType res

Apply an optional parameter to a request

(-&-) :: SwaggerPetstoreRequest req contentType res -> param -> SwaggerPetstoreRequest req contentType res infixl 2

infix operator / alias for addOptionalParam

SwaggerPetstoreRequest

data SwaggerPetstoreRequest req contentType res

Represents a request. The "req" type variable is the request type. The "res" type variable is the response type.

Constructors

SwaggerPetstoreRequest

Fields

rMethod :: Method

Method of SwaggerPetstoreRequest

rUrlPath :: [ByteString]

Endpoint of SwaggerPetstoreRequest

rParams :: Params

params of SwaggerPetstoreRequest

rAuthTypes :: [TypeRep]

types of auth methods

rMethodL :: Lens_' (SwaggerPetstoreRequest req contentType res) Method

rMethod Lens

rUrlPathL :: Lens_' (SwaggerPetstoreRequest req contentType res) [ByteString]

rUrlPath Lens

rParamsL :: Lens_' (SwaggerPetstoreRequest req contentType res) Params

rParams Lens

rAuthTypesL :: Lens_' (SwaggerPetstoreRequest req contentType res) [TypeRep]

rParams Lens

data Params

Request Params

Constructors

Params

Fields

paramsQuery :: Query

paramsHeaders :: RequestHeaders

paramsBody :: ParamBody

paramsQueryL :: Lens_' Params Query

paramsQuery Lens

paramsHeadersL :: Lens_' Params RequestHeaders

paramsHeaders Lens

paramsBodyL :: Lens_' Params ParamBody

paramsBody Lens

data ParamBody

Request Body

Constructors

ParamBodyNone

ParamBodyB ByteString

ParamBodyBL ByteString

ParamBodyFormUrlEncoded Form

ParamBodyMultipartFormData [Part]

SwaggerPetstoreRequest Utils

_mkRequest

Arguments

:: Method
Method

-> [ByteString]
Endpoint

-> SwaggerPetstoreRequest req contentType res
req: Request Type, res: Response Type

_mkParams :: Params

setHeader :: SwaggerPetstoreRequest req contentType res -> [Header] -> SwaggerPetstoreRequest req contentType res

removeHeader :: SwaggerPetstoreRequest req contentType res -> [HeaderName] -> SwaggerPetstoreRequest req contentType res

_setContentTypeHeader :: forall req contentType res. MimeType contentType => SwaggerPetstoreRequest req contentType res -> SwaggerPetstoreRequest req contentType res

_setAcceptHeader :: forall req contentType res accept. MimeType accept => SwaggerPetstoreRequest req contentType res -> accept -> SwaggerPetstoreRequest req contentType res

setQuery :: SwaggerPetstoreRequest req contentType res -> [QueryItem] -> SwaggerPetstoreRequest req contentType res

addForm :: SwaggerPetstoreRequest req contentType res -> Form -> SwaggerPetstoreRequest req contentType res

_addMultiFormPart :: SwaggerPetstoreRequest req contentType res -> Part -> SwaggerPetstoreRequest req contentType res

_setBodyBS :: SwaggerPetstoreRequest req contentType res -> ByteString -> SwaggerPetstoreRequest req contentType res

_setBodyLBS :: SwaggerPetstoreRequest req contentType res -> ByteString -> SwaggerPetstoreRequest req contentType res

_hasAuthType :: AuthMethod authMethod => SwaggerPetstoreRequest req contentType res -> Proxy authMethod -> SwaggerPetstoreRequest req contentType res

Params Utils

toPath :: ToHttpApiData a => a -> ByteString

toHeader :: ToHttpApiData a => (HeaderName, a) -> [Header]

toForm :: ToHttpApiData v => (ByteString, v) -> Form

toQuery :: ToHttpApiData a => (ByteString, Maybe a) -> [QueryItem]

Swagger CollectionFormat Utils

data CollectionFormat

Determines the format of the array if type array is used.

Constructors

CommaSeparated
CSV format for multiple parameters.

SpaceSeparated
Also called SSV

TabSeparated
Also called TSV

PipeSeparated
`value1|value2|value2`

MultiParamArray
Using multiple GET parameters, e.g. `foo=bar&foo=baz`. This is valid only for parameters in "query" (Query) or "formData" (Form)

toHeaderColl :: ToHttpApiData a => CollectionFormat -> (HeaderName, [a]) -> [Header]

toFormColl :: ToHttpApiData v => CollectionFormat -> (ByteString, [v]) -> Form

toQueryColl :: ToHttpApiData a => CollectionFormat -> (ByteString, Maybe [a]) -> Query

_toColl :: Traversable f => CollectionFormat -> (f a -> [(b, ByteString)]) -> f [a] -> [(b, ByteString)]

_toCollA :: (Traversable f, Traversable t, Alternative t) => CollectionFormat -> (f (t a) -> [(b, t ByteString)]) -> f (t [a]) -> [(b, t ByteString)]

_toCollA' :: (Monoid c, Traversable f, Traversable t, Alternative t) => CollectionFormat -> (f (t a) -> [(b, t c)]) -> (Char -> c) -> f (t [a]) -> [(b, t c)]

AuthMethods

class Typeable a => AuthMethod a where

Provides a method to apply auth methods to requests

Minimal complete definition

applyAuthMethod

Methods

applyAuthMethod :: SwaggerPetstoreRequest req contentType res -> a -> SwaggerPetstoreRequest req contentType res

data AnyAuthMethod

An existential wrapper for any AuthMethod

Constructors

AuthMethod a => AnyAuthMethod a

AuthApiKeyApiKey

data AuthApiKeyApiKey

Constructors

AuthApiKeyApiKey Text
secret

AuthApiKeyApiKeyQuery

data AuthApiKeyApiKeyQuery

Constructors

AuthApiKeyApiKeyQuery Text
secret

AuthBasicHttpBasicTest

data AuthBasicHttpBasicTest

Constructors

AuthBasicHttpBasicTest ByteString ByteString
username password

AuthOAuthPetstoreAuth

data AuthOAuthPetstoreAuth

Constructors

AuthOAuthPetstoreAuth Text
secret