SwaggerPetstore.API

Description

Operations

AnotherFake

testSpecialTags

testSpecialTags

Arguments

:: (Consumes TestSpecialTags contentType, MimeRender contentType Client)

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Client
"body" - client model

-> SwaggerPetstoreRequest TestSpecialTags contentType Client accept

PATCH /another-fake/dummy

To test special tags

To test special tags

data TestSpecialTags

Fake

fakeOuterBooleanSerialize

fakeOuterBooleanSerialize

Arguments

:: Consumes FakeOuterBooleanSerialize contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> SwaggerPetstoreRequest FakeOuterBooleanSerialize contentType OuterBoolean accept

POST /fake/outer/boolean

Test serialization of outer boolean types

data FakeOuterBooleanSerialize

fakeOuterCompositeSerialize

fakeOuterCompositeSerialize

Arguments

:: Consumes FakeOuterCompositeSerialize contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> SwaggerPetstoreRequest FakeOuterCompositeSerialize contentType OuterComposite accept

POST /fake/outer/composite

Test serialization of object with outer number type

data FakeOuterCompositeSerialize

fakeOuterNumberSerialize

fakeOuterNumberSerialize

Arguments

:: Consumes FakeOuterNumberSerialize contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> SwaggerPetstoreRequest FakeOuterNumberSerialize contentType OuterNumber accept

POST /fake/outer/number

Test serialization of outer number types

data FakeOuterNumberSerialize

fakeOuterStringSerialize

fakeOuterStringSerialize

Arguments

:: Consumes FakeOuterStringSerialize contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> SwaggerPetstoreRequest FakeOuterStringSerialize contentType OuterString accept

POST /fake/outer/string

Test serialization of outer string types

data FakeOuterStringSerialize

testClientModel

testClientModel

Arguments

:: (Consumes TestClientModel contentType, MimeRender contentType Client)

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Client
"body" - client model

-> SwaggerPetstoreRequest TestClientModel contentType Client accept

PATCH /fake

To test "client" model

To test "client" model

data TestClientModel

testEndpointParameters

testEndpointParameters

Arguments

:: Consumes TestEndpointParameters contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Number
"number" - None

-> ParamDouble
"double" - None

-> PatternWithoutDelimiter
"patternWithoutDelimiter" - None

-> Byte
"byte" - None

-> SwaggerPetstoreRequest TestEndpointParameters contentType res accept

POST /fake

Fake endpoint for testing various parameters 假端點 偽のエンドポイント 가짜 엔드 포인트

Fake endpoint for testing various parameters 假端點 偽のエンドポイント 가짜 엔드 포인트

AuthMethod: AuthBasicHttpBasicTest

Note: Has Produces instances, but no response schema

data TestEndpointParameters

testEnumParameters

testEnumParameters

Arguments

:: Consumes TestEnumParameters contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> SwaggerPetstoreRequest TestEnumParameters contentType res accept

GET /fake

To test enum parameters

To test enum parameters

Note: Has Produces instances, but no response schema

data TestEnumParameters

testInlineAdditionalProperties

testInlineAdditionalProperties

Arguments

:: (Consumes TestInlineAdditionalProperties contentType, MimeRender contentType Value)

=> ContentType contentType
request content-type (MimeType)

-> Value
"param" - request body

-> SwaggerPetstoreRequest TestInlineAdditionalProperties contentType NoContent MimeNoContent

POST /fake/inline-additionalProperties

test inline additionalProperties

data TestInlineAdditionalProperties

testJsonFormData

testJsonFormData

Arguments

:: Consumes TestJsonFormData contentType

=> ContentType contentType
request content-type (MimeType)

-> Param
"param" - field1

-> Param2
"param2" - field2

-> SwaggerPetstoreRequest TestJsonFormData contentType NoContent MimeNoContent

GET /fake/jsonFormData

test json serialization of form data

data TestJsonFormData

FakeClassnameTags123

testClassname

testClassname

Arguments

:: (Consumes TestClassname contentType, MimeRender contentType Client)

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Client
"body" - client model

-> SwaggerPetstoreRequest TestClassname contentType Client accept

PATCH /fake_classname_test

To test class name in snake case

AuthMethod: AuthApiKeyApiKeyQuery

data TestClassname

Pet

addPet

addPet

Arguments

:: (Consumes AddPet contentType, MimeRender contentType Pet)

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Pet
"body" - Pet object that needs to be added to the store

-> SwaggerPetstoreRequest AddPet contentType res accept

POST /pet

Add a new pet to the store

AuthMethod: AuthOAuthPetstoreAuth

Note: Has Produces instances, but no response schema

data AddPet

deletePet

deletePet

Arguments

:: Accept accept
request accept (MimeType)

-> PetId
"petId" - Pet id to delete

-> SwaggerPetstoreRequest DeletePet MimeNoContent res accept

DELETE /pet/{petId}

Deletes a pet

AuthMethod: AuthOAuthPetstoreAuth

Note: Has Produces instances, but no response schema

data DeletePet

findPetsByStatus

findPetsByStatus

Arguments

:: Accept accept
request accept (MimeType)

-> Status
"status" - Status values that need to be considered for filter

-> SwaggerPetstoreRequest FindPetsByStatus MimeNoContent [Pet] accept

GET /pet/findByStatus

Finds Pets by status

Multiple status values can be provided with comma separated strings

AuthMethod: AuthOAuthPetstoreAuth

data FindPetsByStatus

findPetsByTags

findPetsByTags

Arguments

:: Accept accept
request accept (MimeType)

-> Tags
"tags" - Tags to filter by

-> SwaggerPetstoreRequest FindPetsByTags MimeNoContent [Pet] accept

Deprecated:

GET /pet/findByTags

Finds Pets by tags

Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.

AuthMethod: AuthOAuthPetstoreAuth

data FindPetsByTags

getPetById

getPetById

Arguments

:: Accept accept
request accept (MimeType)

-> PetId
"petId" - ID of pet to return

-> SwaggerPetstoreRequest GetPetById MimeNoContent Pet accept

GET /pet/{petId}

Find pet by ID

Returns a single pet

AuthMethod: AuthApiKeyApiKey

data GetPetById

updatePet

updatePet

Arguments

:: (Consumes UpdatePet contentType, MimeRender contentType Pet)

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Pet
"body" - Pet object that needs to be added to the store

-> SwaggerPetstoreRequest UpdatePet contentType res accept

PUT /pet

Update an existing pet

AuthMethod: AuthOAuthPetstoreAuth

Note: Has Produces instances, but no response schema

data UpdatePet

updatePetWithForm

updatePetWithForm

Arguments

:: Consumes UpdatePetWithForm contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> PetId
"petId" - ID of pet that needs to be updated

-> SwaggerPetstoreRequest UpdatePetWithForm contentType res accept

POST /pet/{petId}

Updates a pet in the store with form data

AuthMethod: AuthOAuthPetstoreAuth

Note: Has Produces instances, but no response schema

data UpdatePetWithForm

uploadFile

uploadFile

Arguments

:: Consumes UploadFile contentType

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> PetId
"petId" - ID of pet to update

-> SwaggerPetstoreRequest UploadFile contentType ApiResponse accept

POST /pet/{petId}/uploadImage

uploads an image

AuthMethod: AuthOAuthPetstoreAuth

data UploadFile

Store

deleteOrder

deleteOrder

Arguments

:: Accept accept
request accept (MimeType)

-> OrderIdText
"orderId" - ID of the order that needs to be deleted

-> SwaggerPetstoreRequest DeleteOrder MimeNoContent res accept

DELETE /store/order/{order_id}

Delete purchase order by ID

For valid response try integer IDs with value < 1000. Anything above 1000 or nonintegers will generate API errors

Note: Has Produces instances, but no response schema

data DeleteOrder

getInventory

getInventory

Arguments

:: Accept accept
request accept (MimeType)

-> SwaggerPetstoreRequest GetInventory MimeNoContent (Map String Int) accept

GET /store/inventory

Returns pet inventories by status

Returns a map of status codes to quantities

AuthMethod: AuthApiKeyApiKey

data GetInventory

getOrderById

getOrderById

Arguments

:: Accept accept
request accept (MimeType)

-> OrderId
"orderId" - ID of pet that needs to be fetched

-> SwaggerPetstoreRequest GetOrderById MimeNoContent Order accept

GET /store/order/{order_id}

Find purchase order by ID

For valid response try integer IDs with value 5 or 10. Other values will generated exceptions

data GetOrderById

placeOrder

placeOrder

Arguments

:: (Consumes PlaceOrder contentType, MimeRender contentType Order)

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Order
"body" - order placed for purchasing the pet

-> SwaggerPetstoreRequest PlaceOrder contentType Order accept

POST /store/order

Place an order for a pet

data PlaceOrder

User

createUser

createUser

Arguments

:: (Consumes CreateUser contentType, MimeRender contentType User)

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> User
"body" - Created user object

-> SwaggerPetstoreRequest CreateUser contentType res accept

POST /user

Create user

This can only be done by the logged in user.

Note: Has Produces instances, but no response schema

data CreateUser

createUsersWithArrayInput

createUsersWithArrayInput

Arguments

:: (Consumes CreateUsersWithArrayInput contentType, MimeRender contentType Body)

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Body
"body" - List of user object

-> SwaggerPetstoreRequest CreateUsersWithArrayInput contentType res accept

POST /user/createWithArray

Creates list of users with given input array

Note: Has Produces instances, but no response schema

data CreateUsersWithArrayInput

createUsersWithListInput

createUsersWithListInput

Arguments

:: (Consumes CreateUsersWithListInput contentType, MimeRender contentType Body)

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Body
"body" - List of user object

-> SwaggerPetstoreRequest CreateUsersWithListInput contentType res accept

POST /user/createWithList

Creates list of users with given input array

Note: Has Produces instances, but no response schema

data CreateUsersWithListInput

deleteUser

deleteUser

Arguments

:: Accept accept
request accept (MimeType)

-> Username
"username" - The name that needs to be deleted

-> SwaggerPetstoreRequest DeleteUser MimeNoContent res accept

DELETE /user/{username}

Delete user

This can only be done by the logged in user.

Note: Has Produces instances, but no response schema

data DeleteUser

getUserByName

getUserByName

Arguments

:: Accept accept
request accept (MimeType)

-> Username
"username" - The name that needs to be fetched. Use user1 for testing.

-> SwaggerPetstoreRequest GetUserByName MimeNoContent User accept

GET /user/{username}

Get user by user name

data GetUserByName

loginUser

loginUser

Arguments

:: Accept accept
request accept (MimeType)

-> Username
"username" - The user name for login

-> Password
"password" - The password for login in clear text

-> SwaggerPetstoreRequest LoginUser MimeNoContent Text accept

GET /user/login

Logs user into the system

data LoginUser

logoutUser

logoutUser

Arguments

:: Accept accept
request accept (MimeType)

-> SwaggerPetstoreRequest LogoutUser MimeNoContent res accept

GET /user/logout

Logs out current logged in user session

Note: Has Produces instances, but no response schema

data LogoutUser

updateUser

updateUser

Arguments

:: (Consumes UpdateUser contentType, MimeRender contentType User)

=> ContentType contentType
request content-type (MimeType)

-> Accept accept
request accept (MimeType)

-> Username
"username" - name that need to be deleted

-> User
"body" - Updated user object

-> SwaggerPetstoreRequest UpdateUser contentType res accept

PUT /user/{username}

Updated user

This can only be done by the logged in user.

Note: Has Produces instances, but no response schema

data UpdateUser

Parameter newtypes

newtype AdditionalMetadata

Constructors

AdditionalMetadata

Fields

unAdditionalMetadata :: Text

newtype ApiKey

Constructors

ApiKey

Fields

unApiKey :: Text

newtype Body

Constructors

Body

Fields

unBody :: [User]

newtype Byte

Constructors

Byte

Fields

unByte :: ByteArray

newtype Callback

Constructors

Callback

Fields

unCallback :: Text

newtype EnumFormString

Constructors

EnumFormString

Fields

unEnumFormString :: E'EnumFormString

newtype EnumFormStringArray

Constructors

EnumFormStringArray

Fields

unEnumFormStringArray :: [E'Inner2]

newtype EnumHeaderString

Constructors

EnumHeaderString

Fields

unEnumHeaderString :: E'EnumFormString

newtype EnumHeaderStringArray

Constructors

EnumHeaderStringArray

Fields

unEnumHeaderStringArray :: [E'Inner2]

newtype EnumQueryDouble

Constructors

EnumQueryDouble

Fields

unEnumQueryDouble :: E'EnumNumber

newtype EnumQueryInteger

Constructors

EnumQueryInteger

Fields

unEnumQueryInteger :: E'EnumQueryInteger

newtype EnumQueryString

Constructors

EnumQueryString

Fields

unEnumQueryString :: E'EnumFormString

newtype EnumQueryStringArray

Constructors

EnumQueryStringArray

Fields

unEnumQueryStringArray :: [E'Inner2]

newtype File

Constructors

File

Fields

unFile :: FilePath

newtype Int32

Constructors

Int32

Fields

unInt32 :: Int

newtype Int64

Constructors

Int64

Fields

unInt64 :: Integer

newtype Name2

Constructors

Name2

Fields

unName2 :: Text

newtype Number

Constructors

Number

Fields

unNumber :: Double

newtype OrderId

Constructors

OrderId

Fields

unOrderId :: Integer

newtype OrderIdText

Constructors

OrderIdText

Fields

unOrderIdText :: Text

newtype Param

Constructors

Param

Fields

unParam :: Text

newtype Param2

Constructors

Param2

Fields

unParam2 :: Text

newtype ParamBinary

Constructors

ParamBinary

Fields

unParamBinary :: Binary

newtype ParamDate

Constructors

ParamDate

Fields

unParamDate :: Date

newtype ParamDateTime

Constructors

ParamDateTime

Fields

unParamDateTime :: DateTime

newtype ParamDouble

Constructors

ParamDouble

Fields

unParamDouble :: Double

newtype ParamFloat

Constructors

ParamFloat

Fields

unParamFloat :: Float

newtype ParamInteger

Constructors

ParamInteger

Fields

unParamInteger :: Int

newtype ParamString

Constructors

ParamString

Fields

unParamString :: Text

newtype Password

Constructors

Password

Fields

unPassword :: Text

newtype PatternWithoutDelimiter

Constructors

PatternWithoutDelimiter

Fields

unPatternWithoutDelimiter :: Text

newtype PetId

Constructors

PetId

Fields

unPetId :: Integer

newtype Status

Constructors

Status

Fields

unStatus :: [E'Status2]

newtype StatusText

Constructors

StatusText

Fields

unStatusText :: Text

newtype Tags

Constructors

Tags

Fields

unTags :: [Text]

newtype Username

Constructors

Username

Fields

unUsername :: Text

Auth Methods

AuthApiKeyApiKey

data AuthApiKeyApiKey

Constructors

AuthApiKeyApiKey Text
secret

AuthApiKeyApiKeyQuery

data AuthApiKeyApiKeyQuery

Constructors

AuthApiKeyApiKeyQuery Text
secret

AuthBasicHttpBasicTest

data AuthBasicHttpBasicTest

Constructors

AuthBasicHttpBasicTest ByteString ByteString
username password

AuthOAuthPetstoreAuth

data AuthOAuthPetstoreAuth

Constructors

AuthOAuthPetstoreAuth Text
secret

Custom Mime Types

MimeJsonCharsetutf8

data MimeJsonCharsetutf8

Constructors

MimeJsonCharsetutf8

MimeXmlCharsetutf8

data MimeXmlCharsetutf8

Constructors

MimeXmlCharsetutf8