SwaggerPetstore.Client

Description

Config

data SwaggerPetstoreConfig

Constructors

SwaggerPetstoreConfig

Fields

configHost :: ByteString

configUserAgent :: Text

configExecLoggingT :: ExecLoggingT

configFilterLoggingT :: LogSource -> LogLevel -> Bool

mkConfig :: SwaggerPetstoreConfig

withStdoutLogging :: SwaggerPetstoreConfig -> SwaggerPetstoreConfig

withStderrLogging :: SwaggerPetstoreConfig -> SwaggerPetstoreConfig

withNoLogging :: SwaggerPetstoreConfig -> SwaggerPetstoreConfig

Dispatch

data MimeResponse res

Constructors

MimeResponse

Fields

mimeResponseHttp :: Response ByteString

mimeResponseResult :: Either SwaggerPetstoreError res

dispatchReq

Arguments

:: (Produces req accept, MimeUnrender accept res, MimeType contentType)

=> Manager
http-client Connection manager

-> accept
"accept" MimeType

-> SwaggerPetstoreConfig
config

-> SwaggerPetstoreRequest req contentType res
request

-> IO (MimeResponse res)
response

returns both the underlying http response and the parsed result (MimeResponse)

dispatchReqRes

Arguments

:: (Produces req accept, MimeUnrender accept res, MimeType contentType)

=> Manager
http-client Connection manager

-> accept
"accept" MimeType

-> SwaggerPetstoreConfig
config

-> SwaggerPetstoreRequest req contentType res
request

-> IO (Either SwaggerPetstoreError res)
response

like dispatchReq, but only returns the parsed result

dispatchReqLbs

Arguments

:: (Produces req accept, MimeType contentType)

=> Manager
http-client Connection manager

-> accept
"accept" MimeType

-> SwaggerPetstoreConfig
config

-> SwaggerPetstoreRequest req contentType res
request

-> IO (Response ByteString)
response

like dispatchReq, but only returns the underlying http response

dispatchReqLbsUnsafe

Arguments

:: (MimeType accept, MimeType contentType)

=> Manager
http-client Connection manager

-> accept
"accept" MimeType

-> SwaggerPetstoreConfig
config

-> SwaggerPetstoreRequest req contentType res
request

-> IO (Response ByteString)
response

like dispatchReqLbs, but does not validate the operation is a Producer of the "accept" MimeType. (Useful if the server's response is undocumented)

dispatchReqLbsUnsafeRaw

Arguments

:: MimeType contentType

=> Manager
http-client Connection manager

-> SwaggerPetstoreConfig
config

-> SwaggerPetstoreRequest req contentType res
request

-> IO (Response ByteString)
response

like dispatchReqLbsUnsafe, but does not add an "accept" header.

dispatchInitLbsUnsafe

Arguments

:: Manager
http-client Connection manager

-> SwaggerPetstoreConfig
config

-> InitRequest req contentType res accept
init request

-> IO (Response ByteString)
response

dispatch an InitRequest

InitRequest

newtype InitRequest req contentType res accept

wraps an http-client Request with request/response type parameters

Constructors

InitRequest

Fields

unInitRequest :: Request

_toInitRequest

Arguments

:: (MimeType accept, MimeType contentType)

=> accept
"accept" MimeType

-> SwaggerPetstoreConfig
config

-> SwaggerPetstoreRequest req contentType res
request

-> IO (InitRequest req contentType res accept)
initialized request

Build an http-client Request record from the supplied config and request

modifyInitRequest :: InitRequest req contentType res accept -> (Request -> Request) -> InitRequest req contentType res accept

convenience method for modifying the underlying Request

modifyInitRequestM :: Monad m => InitRequest req contentType res accept -> (Request -> m Request) -> m (InitRequest req contentType res accept)

convenience method for modifying the underlying Request (monadic)

Error

data SwaggerPetstoreError

Constructors

SwaggerPetstoreError

Fields

parseError :: String

reponseError :: Response ByteString

Logging

type ExecLoggingT = forall m. MonadIO m => forall a. LoggingT m a -> m a

runs the logger

Null Logger

nullLogger :: Loc -> LogSource -> LogLevel -> LogStr -> IO ()

runNullLoggingT :: LoggingT m a -> m a

Logging Filters

errorLevelFilter :: LogSource -> LogLevel -> Bool

infoLevelFilter :: LogSource -> LogLevel -> Bool

debugLevelFilter :: LogSource -> LogLevel -> Bool

minLevelFilter :: LogLevel -> LogSource -> LogLevel -> Bool