SwaggerPetstore.Client

Description

Dispatch

Lbs

dispatchLbs

Arguments

:: (Produces req accept, MimeType contentType)

=> Manager
http-client Connection manager

-> SwaggerPetstoreConfig
config

-> SwaggerPetstoreRequest req contentType res accept
request

-> IO (Response ByteString)
response

send a request returning the raw http response

Mime

data MimeResult res

pair of decoded http body and http response

Constructors

MimeResult

Fields

mimeResult :: Either MimeError res

decoded http body

mimeResultResponse :: Response ByteString

http response

data MimeError

pair of unrender/parser error and http response

Constructors

MimeError

Fields

mimeError :: String

unrender/parser error

mimeErrorResponse :: Response ByteString

http response

dispatchMime

Arguments

:: (Produces req accept, MimeUnrender accept res, MimeType contentType)

=> Manager
http-client Connection manager

-> SwaggerPetstoreConfig
config

-> SwaggerPetstoreRequest req contentType res accept
request

-> IO (MimeResult res)
response

send a request returning the MimeResult

dispatchMime'

Arguments

:: (Produces req accept, MimeUnrender accept res, MimeType contentType)

=> Manager
http-client Connection manager

-> SwaggerPetstoreConfig
config

-> SwaggerPetstoreRequest req contentType res accept
request

-> IO (Either MimeError res)
response

like dispatchMime, but only returns the decoded http body

Unsafe

dispatchLbsUnsafe

Arguments

:: (MimeType accept, MimeType contentType)

=> Manager
http-client Connection manager

-> SwaggerPetstoreConfig
config

-> SwaggerPetstoreRequest req contentType res accept
request

-> IO (Response ByteString)
response

like dispatchReqLbs, but does not validate the operation is a Producer of the "accept" MimeType. (Useful if the server's response is undocumented)

dispatchInitUnsafe

Arguments

:: Manager
http-client Connection manager

-> SwaggerPetstoreConfig
config

-> InitRequest req contentType res accept
init request

-> IO (Response ByteString)
response

dispatch an InitRequest

InitRequest

newtype InitRequest req contentType res accept

wraps an http-client Request with request/response type parameters

Constructors

InitRequest

Fields

unInitRequest :: Request

_toInitRequest

Arguments

:: (MimeType accept, MimeType contentType)

=> SwaggerPetstoreConfig
config

-> SwaggerPetstoreRequest req contentType res accept
request

-> IO (InitRequest req contentType res accept)
initialized request

Build an http-client Request record from the supplied config and request

modifyInitRequest :: InitRequest req contentType res accept -> (Request -> Request) -> InitRequest req contentType res accept

modify the underlying Request

modifyInitRequestM :: Monad m => InitRequest req contentType res accept -> (Request -> m Request) -> m (InitRequest req contentType res accept)

modify the underlying Request (monadic)

Logging

runConfigLog :: MonadIO m => SwaggerPetstoreConfig -> LogExec m

Run a block using the configured logger instance

runConfigLogWithExceptions :: (MonadCatch m, MonadIO m) => Text -> SwaggerPetstoreConfig -> LogExec m

Run a block using the configured logger instance (logs exceptions)