SwaggerPetstore.Model

Description

Models

AdditionalPropertiesClass

data AdditionalPropertiesClass

AdditionalPropertiesClass

Constructors

AdditionalPropertiesClass

Fields

additionalPropertiesClassMapProperty :: !(Maybe (Map String Text))

"map_property"

additionalPropertiesClassMapOfMapProperty :: !(Maybe (Map String (Map String Text)))

"map_of_map_property"

mkAdditionalPropertiesClass :: AdditionalPropertiesClass

Construct a value of type AdditionalPropertiesClass (by applying it's required fields, if any)

Animal

data Animal

Animal

Constructors

Animal

Fields

animalClassName :: !Text

Required "className"

animalColor :: !(Maybe Text)

"color"

mkAnimal

Arguments

:: Text
animalClassName

-> Animal

Construct a value of type Animal (by applying it's required fields, if any)

AnimalFarm

data AnimalFarm

AnimalFarm

Constructors

AnimalFarm

mkAnimalFarm :: AnimalFarm

Construct a value of type AnimalFarm (by applying it's required fields, if any)

ApiResponse

data ApiResponse

ApiResponse

Constructors

ApiResponse

Fields

apiResponseCode :: !(Maybe Int)

"code"

apiResponseType :: !(Maybe Text)

"type"

apiResponseMessage :: !(Maybe Text)

"message"

mkApiResponse :: ApiResponse

Construct a value of type ApiResponse (by applying it's required fields, if any)

ArrayOfArrayOfNumberOnly

data ArrayOfArrayOfNumberOnly

ArrayOfArrayOfNumberOnly

Constructors

ArrayOfArrayOfNumberOnly

Fields

arrayOfArrayOfNumberOnlyArrayArrayNumber :: !(Maybe [[Double]])

ArrayArrayNumber

mkArrayOfArrayOfNumberOnly :: ArrayOfArrayOfNumberOnly

Construct a value of type ArrayOfArrayOfNumberOnly (by applying it's required fields, if any)

ArrayOfNumberOnly

data ArrayOfNumberOnly

ArrayOfNumberOnly

Constructors

ArrayOfNumberOnly

Fields

arrayOfNumberOnlyArrayNumber :: !(Maybe [Double])

ArrayNumber

mkArrayOfNumberOnly :: ArrayOfNumberOnly

Construct a value of type ArrayOfNumberOnly (by applying it's required fields, if any)

ArrayTest

data ArrayTest

ArrayTest

Constructors

ArrayTest

Fields

arrayTestArrayOfString :: !(Maybe [Text])

"array_of_string"

arrayTestArrayArrayOfInteger :: !(Maybe [[Integer]])

"array_array_of_integer"

arrayTestArrayArrayOfModel :: !(Maybe [[ReadOnlyFirst]])

"array_array_of_model"

mkArrayTest :: ArrayTest

Construct a value of type ArrayTest (by applying it's required fields, if any)

Capitalization

data Capitalization

Capitalization

Constructors

Capitalization

Fields

capitalizationSmallCamel :: !(Maybe Text)

"smallCamel"

capitalizationCapitalCamel :: !(Maybe Text)

CapitalCamel

capitalizationSmallSnake :: !(Maybe Text)

"small_Snake"

capitalizationCapitalSnake :: !(Maybe Text)

Capital_Snake

capitalizationScaEthFlowPoints :: !(Maybe Text)

SCA_ETH_Flow_Points

capitalizationAttName :: !(Maybe Text)

ATT_NAME - Name of the pet

mkCapitalization :: Capitalization

Construct a value of type Capitalization (by applying it's required fields, if any)

Category

data Category

Category

Constructors

Category

Fields

categoryId :: !(Maybe Integer)

"id"

categoryName :: !(Maybe Text)

"name"

mkCategory :: Category

Construct a value of type Category (by applying it's required fields, if any)

ClassModel

data ClassModel

ClassModel Model for testing model with "_class" property

Constructors

ClassModel

Fields

classModelClass :: !(Maybe Text)

"_class"

mkClassModel :: ClassModel

Construct a value of type ClassModel (by applying it's required fields, if any)

Client

data Client

Client

Constructors

Client

Fields

clientClient :: !(Maybe Text)

"client"

mkClient :: Client

Construct a value of type Client (by applying it's required fields, if any)

EnumArrays

data EnumArrays

EnumArrays

Constructors

EnumArrays

Fields

enumArraysJustSymbol :: !(Maybe E'JustSymbol)

"just_symbol"

enumArraysArrayEnum :: !(Maybe [E'ArrayEnum])

"array_enum"

mkEnumArrays :: EnumArrays

Construct a value of type EnumArrays (by applying it's required fields, if any)

EnumTest

data EnumTest

EnumTest

Constructors

EnumTest

Fields

enumTestEnumString :: !(Maybe E'EnumString)

"enum_string"

enumTestEnumInteger :: !(Maybe E'EnumInteger)

"enum_integer"

enumTestEnumNumber :: !(Maybe E'EnumNumber)

"enum_number"

enumTestOuterEnum :: !(Maybe OuterEnum)

"outerEnum"

mkEnumTest :: EnumTest

Construct a value of type EnumTest (by applying it's required fields, if any)

FormatTest

data FormatTest

FormatTest

Constructors

FormatTest

Fields

formatTestInteger :: !(Maybe Int)

"integer"

formatTestInt32 :: !(Maybe Int)

"int32"

formatTestInt64 :: !(Maybe Integer)

"int64"

formatTestNumber :: !Double

Required "number"

formatTestFloat :: !(Maybe Float)

"float"

formatTestDouble :: !(Maybe Double)

"double"

formatTestString :: !(Maybe Text)

"string"

formatTestByte :: !ByteArray

Required "byte"

formatTestBinary :: !(Maybe Binary)

"binary"

formatTestDate :: !Date

Required "date"

formatTestDateTime :: !(Maybe DateTime)

"dateTime"

formatTestUuid :: !(Maybe Text)

"uuid"

formatTestPassword :: !Text

Required "password"

mkFormatTest

Arguments

:: Double
formatTestNumber

-> ByteArray
formatTestByte

-> Date
formatTestDate

-> Text
formatTestPassword

-> FormatTest

Construct a value of type FormatTest (by applying it's required fields, if any)

HasOnlyReadOnly

data HasOnlyReadOnly

HasOnlyReadOnly

Constructors

HasOnlyReadOnly

Fields

hasOnlyReadOnlyBar :: !(Maybe Text)

"bar"

hasOnlyReadOnlyFoo :: !(Maybe Text)

"foo"

mkHasOnlyReadOnly :: HasOnlyReadOnly

Construct a value of type HasOnlyReadOnly (by applying it's required fields, if any)

MapTest

data MapTest

MapTest

Constructors

MapTest

Fields

mapTestMapMapOfString :: !(Maybe (Map String (Map String Text)))

"map_map_of_string"

mapTestMapOfEnumString :: !(Maybe (Map String E'Inner))

"map_of_enum_string"

mkMapTest :: MapTest

Construct a value of type MapTest (by applying it's required fields, if any)

MixedPropertiesAndAdditionalPropertiesClass

data MixedPropertiesAndAdditionalPropertiesClass

MixedPropertiesAndAdditionalPropertiesClass

Constructors

MixedPropertiesAndAdditionalPropertiesClass

Fields

mixedPropertiesAndAdditionalPropertiesClassUuid :: !(Maybe Text)

"uuid"

mixedPropertiesAndAdditionalPropertiesClassDateTime :: !(Maybe DateTime)

"dateTime"

mixedPropertiesAndAdditionalPropertiesClassMap :: !(Maybe (Map String Animal))

"map"

mkMixedPropertiesAndAdditionalPropertiesClass :: MixedPropertiesAndAdditionalPropertiesClass

Construct a value of type MixedPropertiesAndAdditionalPropertiesClass (by applying it's required fields, if any)

Model200Response

data Model200Response

Model200Response Model for testing model name starting with number

Constructors

Model200Response

Fields

model200ResponseName :: !(Maybe Int)

"name"

model200ResponseClass :: !(Maybe Text)

"class"

mkModel200Response :: Model200Response

Construct a value of type Model200Response (by applying it's required fields, if any)

ModelList

data ModelList

ModelList

Constructors

ModelList

Fields

modelList123List :: !(Maybe Text)

"123-list"

mkModelList :: ModelList

Construct a value of type ModelList (by applying it's required fields, if any)

ModelReturn

data ModelReturn

ModelReturn Model for testing reserved words

Constructors

ModelReturn

Fields

modelReturnReturn :: !(Maybe Int)

"return"

mkModelReturn :: ModelReturn

Construct a value of type ModelReturn (by applying it's required fields, if any)

Name

data Name

Name Model for testing model name same as property name

Constructors

Name

Fields

nameName :: !Int

Required "name"

nameSnakeCase :: !(Maybe Int)

"snake_case"

nameProperty :: !(Maybe Text)

"property"

name123Number :: !(Maybe Int)

"123Number"

mkName

Arguments

:: Int
nameName

-> Name

Construct a value of type Name (by applying it's required fields, if any)

NumberOnly

data NumberOnly

NumberOnly

Constructors

NumberOnly

Fields

numberOnlyJustNumber :: !(Maybe Double)

JustNumber

mkNumberOnly :: NumberOnly

Construct a value of type NumberOnly (by applying it's required fields, if any)

Order

data Order

Order

Constructors

Order

Fields

orderId :: !(Maybe Integer)

"id"

orderPetId :: !(Maybe Integer)

"petId"

orderQuantity :: !(Maybe Int)

"quantity"

orderShipDate :: !(Maybe DateTime)

"shipDate"

orderStatus :: !(Maybe E'Status)

"status" - Order Status

orderComplete :: !(Maybe Bool)

"complete"

mkOrder :: Order

Construct a value of type Order (by applying it's required fields, if any)

OuterBoolean

newtype OuterBoolean

OuterBoolean

Constructors

OuterBoolean

Fields

unOuterBoolean :: Bool

OuterComposite

data OuterComposite

OuterComposite

Constructors

OuterComposite

Fields

outerCompositeMyNumber :: !(Maybe OuterNumber)

"my_number"

outerCompositeMyString :: !(Maybe OuterString)

"my_string"

outerCompositeMyBoolean :: !(Maybe OuterBoolean)

"my_boolean"

mkOuterComposite :: OuterComposite

Construct a value of type OuterComposite (by applying it's required fields, if any)

OuterNumber

newtype OuterNumber

OuterNumber

Constructors

OuterNumber

Fields

unOuterNumber :: Double

OuterString

newtype OuterString

OuterString

Constructors

OuterString

Fields

unOuterString :: Text

Pet

data Pet

Pet

Constructors

Pet

Fields

petId :: !(Maybe Integer)

"id"

petCategory :: !(Maybe Category)

"category"

petName :: !Text

Required "name"

petPhotoUrls :: ![Text]

Required "photoUrls"

petTags :: !(Maybe [Tag])

"tags"

petStatus :: !(Maybe E'Status2)

"status" - pet status in the store

mkPet

Arguments

:: Text
petName

-> [Text]
petPhotoUrls

-> Pet

Construct a value of type Pet (by applying it's required fields, if any)

ReadOnlyFirst

data ReadOnlyFirst

ReadOnlyFirst

Constructors

ReadOnlyFirst

Fields

readOnlyFirstBar :: !(Maybe Text)

"bar"

readOnlyFirstBaz :: !(Maybe Text)

"baz"

mkReadOnlyFirst :: ReadOnlyFirst

Construct a value of type ReadOnlyFirst (by applying it's required fields, if any)

SpecialModelName

data SpecialModelName

SpecialModelName

Constructors

SpecialModelName

Fields

specialModelNameSpecialPropertyName :: !(Maybe Integer)

"$special[property.name]"

mkSpecialModelName :: SpecialModelName

Construct a value of type SpecialModelName (by applying it's required fields, if any)

Tag

data Tag

Tag

Constructors

Tag

Fields

tagId :: !(Maybe Integer)

"id"

tagName :: !(Maybe Text)

"name"

mkTag :: Tag

Construct a value of type Tag (by applying it's required fields, if any)

User

data User

User

Constructors

User

Fields

userId :: !(Maybe Integer)

"id"

userUsername :: !(Maybe Text)

"username"

userFirstName :: !(Maybe Text)

"firstName"

userLastName :: !(Maybe Text)

"lastName"

userEmail :: !(Maybe Text)

"email"

userPassword :: !(Maybe Text)

"password"

userPhone :: !(Maybe Text)

"phone"

userUserStatus :: !(Maybe Int)

"userStatus" - User Status

mkUser :: User

Construct a value of type User (by applying it's required fields, if any)

Cat

data Cat

Cat

Constructors

Cat

Fields

catClassName :: !Text

Required "className"

catColor :: !(Maybe Text)

"color"

catDeclawed :: !(Maybe Bool)

"declawed"

mkCat

Arguments

:: Text
catClassName

-> Cat

Construct a value of type Cat (by applying it's required fields, if any)

Dog

data Dog

Dog

Constructors

Dog

Fields

dogClassName :: !Text

Required "className"

dogColor :: !(Maybe Text)

"color"

dogBreed :: !(Maybe Text)

"breed"

mkDog

Arguments

:: Text
dogClassName

-> Dog

Construct a value of type Dog (by applying it's required fields, if any)

Enums

E'ArrayEnum

data E'ArrayEnum

Enum of Text

Constructors

E'ArrayEnum'Fish
"fish"

E'ArrayEnum'Crab
"crab"

fromE'ArrayEnum :: E'ArrayEnum -> Text

unwrap E'ArrayEnum enum

toE'ArrayEnum :: Text -> Either String E'ArrayEnum

parse E'ArrayEnum enum

E'EnumFormString

data E'EnumFormString

Enum of Text

Constructors

E'EnumFormString'_abc
"_abc"

E'EnumFormString'_efg
"-efg"

E'EnumFormString'_xyz
"(xyz)"

fromE'EnumFormString :: E'EnumFormString -> Text

unwrap E'EnumFormString enum

toE'EnumFormString :: Text -> Either String E'EnumFormString

parse E'EnumFormString enum

E'EnumInteger

data E'EnumInteger

Enum of Int

Constructors

E'EnumInteger'Num1
1

E'EnumInteger'NumMinus_1
-1

fromE'EnumInteger :: E'EnumInteger -> Int

unwrap E'EnumInteger enum

toE'EnumInteger :: Int -> Either String E'EnumInteger

parse E'EnumInteger enum

E'EnumNumber

data E'EnumNumber

Enum of Double

Constructors

E'EnumNumber'Num1_Dot_1
1.1

E'EnumNumber'NumMinus_1_Dot_2
-1.2

fromE'EnumNumber :: E'EnumNumber -> Double

unwrap E'EnumNumber enum

toE'EnumNumber :: Double -> Either String E'EnumNumber

parse E'EnumNumber enum

E'EnumQueryInteger

data E'EnumQueryInteger

Enum of Int

Constructors

E'EnumQueryInteger'Num1
1

E'EnumQueryInteger'NumMinus_2
-2

fromE'EnumQueryInteger :: E'EnumQueryInteger -> Int

unwrap E'EnumQueryInteger enum

toE'EnumQueryInteger :: Int -> Either String E'EnumQueryInteger

parse E'EnumQueryInteger enum

E'EnumString

data E'EnumString

Enum of Text

Constructors

E'EnumString'UPPER
UPPER

E'EnumString'Lower
"lower"

E'EnumString'Empty
""

fromE'EnumString :: E'EnumString -> Text

unwrap E'EnumString enum

toE'EnumString :: Text -> Either String E'EnumString

parse E'EnumString enum

E'Inner

data E'Inner

Enum of Text

Constructors

E'Inner'UPPER
UPPER

E'Inner'Lower
"lower"

fromE'Inner :: E'Inner -> Text

unwrap E'Inner enum

toE'Inner :: Text -> Either String E'Inner

parse E'Inner enum

E'Inner2

data E'Inner2

Enum of Text

Constructors

E'Inner2'GreaterThan
">"

E'Inner2'Dollar
"$"

fromE'Inner2 :: E'Inner2 -> Text

unwrap E'Inner2 enum

toE'Inner2 :: Text -> Either String E'Inner2

parse E'Inner2 enum

E'JustSymbol

data E'JustSymbol

Enum of Text

Constructors

E'JustSymbol'Greater_Than_Or_Equal_To
">="

E'JustSymbol'Dollar
"$"

fromE'JustSymbol :: E'JustSymbol -> Text

unwrap E'JustSymbol enum

toE'JustSymbol :: Text -> Either String E'JustSymbol

parse E'JustSymbol enum

E'Status

data E'Status

Enum of Text . Order Status

Constructors

E'Status'Placed
"placed"

E'Status'Approved
"approved"

E'Status'Delivered
"delivered"

fromE'Status :: E'Status -> Text

unwrap E'Status enum

toE'Status :: Text -> Either String E'Status

parse E'Status enum

E'Status2

data E'Status2

Enum of Text . pet status in the store

Constructors

E'Status2'Available
"available"

E'Status2'Pending
"pending"

E'Status2'Sold
"sold"

fromE'Status2 :: E'Status2 -> Text

unwrap E'Status2 enum

toE'Status2 :: Text -> Either String E'Status2

parse E'Status2 enum

EnumClass

data EnumClass

Enum of Text

Constructors

EnumClass'_abc
"_abc"

EnumClass'_efg
"-efg"

EnumClass'_xyz
"(xyz)"

fromEnumClass :: EnumClass -> Text

unwrap EnumClass enum

toEnumClass :: Text -> Either String EnumClass

parse EnumClass enum

OuterEnum

data OuterEnum

Enum of Text

Constructors

OuterEnum'Placed
"placed"

OuterEnum'Approved
"approved"

OuterEnum'Delivered
"delivered"

fromOuterEnum :: OuterEnum -> Text

unwrap OuterEnum enum

toOuterEnum :: Text -> Either String OuterEnum

parse OuterEnum enum