SwaggerPetstore.Model

Description

Models

ApiResponse

data ApiResponse

An uploaded response

Describes the result of uploading an image resource

Constructors

ApiResponse

Fields

apiResponseCode :: !(Maybe Int)

"code"

apiResponseType :: !(Maybe Text)

"type"

apiResponseMessage :: !(Maybe Text)

"message"

mkApiResponse :: ApiResponse

Construct a value of type ApiResponse (by applying it's required fields, if any)

Category

data Category

Pet catehgry

A category for a pet

Constructors

Category

Fields

categoryId :: !(Maybe Integer)

"id"

categoryName :: !(Maybe Text)

"name"

mkCategory :: Category

Construct a value of type Category (by applying it's required fields, if any)

Order

data Order

Pet Order

An order for a pets from the pet store

Constructors

Order

Fields

orderId :: !(Maybe Integer)

"id"

orderPetId :: !(Maybe Integer)

"petId"

orderQuantity :: !(Maybe Int)

"quantity"

orderShipDate :: !(Maybe UTCTime)

"shipDate"

orderStatus :: !(Maybe Text)

"status" - Order Status

orderComplete :: !(Maybe Bool)

"complete"

mkOrder :: Order

Construct a value of type Order (by applying it's required fields, if any)

Pet

data Pet

a Pet

A pet for sale in the pet store

Constructors

Pet

Fields

petId :: !(Maybe Integer)

"id"

petCategory :: !(Maybe Category)

"category"

petName :: !Text

Required "name"

petPhotoUrls :: ![Text]

Required "photoUrls"

petTags :: !(Maybe [Tag])

"tags"

petStatus :: !(Maybe Text)

"status" - pet status in the store

mkPet

Arguments

:: Text
petName

-> [Text]
petPhotoUrls

-> Pet

Construct a value of type Pet (by applying it's required fields, if any)

Tag

data Tag

Pet Tag

A tag for a pet

Constructors

Tag

Fields

tagId :: !(Maybe Integer)

"id"

tagName :: !(Maybe Text)

"name"

mkTag :: Tag

Construct a value of type Tag (by applying it's required fields, if any)

User

data User

a User

A User who is purchasing from the pet store

Constructors

User

Fields

userId :: !(Maybe Integer)

"id"

userUsername :: !(Maybe Text)

"username"

userFirstName :: !(Maybe Text)

"firstName"

userLastName :: !(Maybe Text)

"lastName"

userEmail :: !(Maybe Text)

"email"

userPassword :: !(Maybe Text)

"password"

userPhone :: !(Maybe Text)

"phone"

userUserStatus :: !(Maybe Int)

"userStatus" - User Status

mkUser :: User

Construct a value of type User (by applying it's required fields, if any)

Utils

_omitNulls :: [(Text, Value)] -> Value

Removes Null fields. (OpenAPI-Specification 2.0 does not allow Null in JSON)

_toFormItem :: (ToHttpApiData a, Functor f) => t -> f a -> f (t, [Text])

_emptyToNothing :: Maybe String -> Maybe String

_memptyToNothing :: (Monoid a, Eq a) => Maybe a -> Maybe a

DateTime Formatting

_readDateTime :: (ParseTime t, Monad m, Alternative m) => String -> m t

_parseISO8601

_showDateTime :: (t ~ UTCTime, FormatTime t) => t -> String

TI.formatISO8601Millis

_parseISO8601 :: (ParseTime t, Monad m, Alternative m) => String -> m t

Date Formatting

_readDate :: (ParseTime t, Monad m) => String -> m t

TI.parseTimeM True TI.defaultTimeLocale "%Y-%m-%d"

_showDate :: FormatTime t => t -> String

TI.formatTime TI.defaultTimeLocale "%Y-%m-%d"